### 2025-06-14 47565d3
docs: fix author affiliations and bib DOIs
