+++ joss/paper.bib
@@ -16,6 +16,7 @@
   journal={Systematic Biology},
   pages={syae030},
   year={2024},
+  doi={10.1093/sysbio/syae030},
   publisher={Oxford University Press US}
 }
 
@@ -27,6 +28,7 @@
     year = {2019},
     volume = {35},
     pages = {526-528},
+    doi = {10.1093/bioinformatics/bty633},
 }
 
 @article{suchard2009,
@@ -37,6 +39,7 @@
   number={11},
   pages={1370--1376},
   year={2009},
+  doi={10.1093/bioinformatics/btp244},
   publisher={Oxford University Press}
 }
 
@@ -48,16 +51,17 @@
   number={1},
   pages={170--173},
   year={2012},
+  doi={10.1093/sysbio/syr100},
   publisher={Oxford University Press}
 }
 
 @article{adelson1962,
   title={An algorithm for the organization of information},
   author={Adelson-Velsky, Georgii, and Landis, Evgenii},
-  journal={Soviet Math.},
-  volume={3},
-  pages={1259--1263},
-  year={1962}
+  journal={Proc. USSR Acad. Sci.},
+  volume={146},
+  pages={263--266},
+  year={In Russian. English translation in Soviet Mathematics Doklady, 3:1259--1263, 1962}
 }
 
 @article{penn2023,
@@ -68,6 +72,7 @@
   number={12},
   pages={evad213},
   year={2023},
+  doi={10.1093/gbe/evad213},
   publisher={Oxford University Press US}
 }
 
@@ -92,6 +97,7 @@
   number={12},
   pages={719--728},
   year={2013},
+  doi={10.1016/j.tree.2013.09.004},
   publisher={Elsevier}
 }
 
@@ -103,6 +109,7 @@
   number={3},
   pages={444--446},
   year={2016},
+  doi={10.1093/bioinformatics/btv573},
   publisher={Oxford University Press}
 }
 
@@ -114,6 +121,7 @@
   number={6},
   pages={1635--1638},
   year={2016},
+  doi={10.1093/molbev/msw046},
   publisher={Society for Molecular Biology and Evolution}
 }
 
@@ -126,7 +134,7 @@
   number = {101},
   pages = {6943},
   author = {Matthew Andres Moreno and Mark T. Holder and Jeet Sukumaran},
-  title = {DendroPy 5: a mature Python library for phylogenetic computing},
+  title = {{DendroPy 5: a mature Python library for phylogenetic computing}},
   journal = {Journal of Open Source Software}
 }
 
@@ -138,15 +146,17 @@
   number={7},
   pages={428--444},
   year={2020},
+  doi={10.1038/s41576-020-0233-0},
   publisher={Nature Publishing Group UK London}
 }
 
 @article{sanderson2022,
   title={Taxonium, a web-based tool for exploring large phylogenetic trees},
   author={Sanderson, Theo},
-  journal={Elife},
+  journal={eLife},
   volume={11},
   year={2022},
+  doi={10.7554/eLife.82392},
   publisher={eLife Sciences Publications, Ltd}
 }
 
@@ -158,6 +168,7 @@
   number={5},
   pages={746--752},
   year={2023},
+  doi={10.1038/s41588-023-01368-0},
   publisher={Nature Publishing Group US New York}
 }
 
@@ -169,6 +180,7 @@
   number={9},
   pages={547--562},
   year={2022},
+  doi={10.1038/s41576-022-00483-8},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -178,6 +190,7 @@
   journal={bioRxiv},
   pages={2021--10},
   year={2021},
+  doi={10.1101/2021.10.27.465994},
[…]
