### 2025-06-14 192afd8
docs(paper): add country for affil 1
