### 2025-06-14 bc48576
paper: fix dois and affiliations (#76)
* docs: fix author affiliations and bib DOIs

* docs(paper): add country for affil 1
