+++ .github/workflows/ci-R.yml
@@ -20,7 +20,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: prefix-dev/setup-pixi@v0.8.8
+      - uses: prefix-dev/setup-pixi@v0.8.10
         with:
           environments: ${{ env.R_ENV }}
       - run: pixi run -e ${{ env.R_ENV }} test "./${{ env.R_ENV}}"
