### 2025-06-25 61403c9
feat: add benchmarks for base matrix operations
