### 2025-06-25 b379746
chore: add matrix benchmark tests in python
