+++ py-phylo2vec/benchmarks/test_bench.py
@@ -8,7 +8,7 @@ BIG_RANGE = range(10000, 110000, 10000)
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_to_newick_ordered(benchmark, sample_size):
+def test_vector_to_newick(benchmark, sample_size):
     """Benchmark the conversion of an ordered vector to Newick format.
 
     Parameters
@@ -22,7 +22,7 @@ def test_to_newick_ordered(benchmark, sample_size):
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_to_newick_unordered(benchmark, sample_size):
+def test_vector_to_newick_unordered(benchmark, sample_size):
     """Benchmark the conversion of an unordered vector to Newick format.
 
     Parameters
@@ -36,7 +36,21 @@ def test_to_newick_unordered(benchmark, sample_size):
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_to_vector(benchmark, sample_size):
+def test_matrix_from_newick(benchmark, sample_size):
+    """Benchmark the conversion of an unordered matrix to Newick format.
+
+    Parameters
+    ----------
+    benchmark : _pytest.fixtures.BenchmarkFixture
+        Benchmark fixture from pytest-benchmark
+    sample_size : int
+        Number of leaves in the vector
+    """
+    benchmark(core.to_newick_from_matrix, core.sample_matrix(sample_size, False))
+
+
+@pytest.mark.parametrize("sample_size", BIG_RANGE)
+def test_vector_from_newick(benchmark, sample_size):
     """Benchmark the conversion of a Newick string (without branch lengths) to vector format
 
     Parameters
@@ -48,5 +62,22 @@ def test_to_vector(benchmark, sample_size):
     """
     benchmark(
         core.to_vector,
-        core.to_newick_from_vector(core.sample_vector(sample_size, True)),
+        core.to_newick_from_vector(core.sample_vector(sample_size, False)),
+    )
+
+
+@pytest.mark.parametrize("sample_size", BIG_RANGE)
+def test_matrix_to_newick(benchmark, sample_size):
+    """Benchmark the conversion of a Newick string (with branch lengths) to a matrix format
+
+    Parameters
+    ----------
+    benchmark : _pytest.fixtures.BenchmarkFixture
+        Benchmark fixture from pytest-benchmark
+    sample_size : int
+        Number of leaves in the vector
+    """
+    benchmark(
+        core.to_matrix,
+        core.to_newick_from_matrix(core.sample_matrix(sample_size, False)),
     )
