### 2025-06-25 c7fa3bc
fix (py): fix benchmark test names
