+++ py-phylo2vec/benchmarks/test_bench.py
@@ -8,7 +8,7 @@ BIG_RANGE = range(10000, 110000, 10000)
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_vector_to_newick(benchmark, sample_size):
+def test_vector_to_newick_ordered(benchmark, sample_size):
     """Benchmark the conversion of an ordered vector to Newick format.
 
     Parameters
@@ -22,7 +22,7 @@ def test_vector_to_newick(benchmark, sample_size):
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_vector_to_newick_unordered(benchmark, sample_size):
+def test_vector_to_newick(benchmark, sample_size):
     """Benchmark the conversion of an unordered vector to Newick format.
 
     Parameters
@@ -33,20 +33,6 @@ def test_vector_to_newick_unordered(benchmark, sample_size):
         Number of leaves in the vector
     """
     benchmark(core.to_newick_from_vector, core.sample_vector(sample_size, False))
-
-
-@pytest.mark.parametrize("sample_size", BIG_RANGE)
-def test_matrix_from_newick(benchmark, sample_size):
-    """Benchmark the conversion of an unordered matrix to Newick format.
-
-    Parameters
-    ----------
-    benchmark : _pytest.fixtures.BenchmarkFixture
-        Benchmark fixture from pytest-benchmark
-    sample_size : int
-        Number of leaves in the vector
-    """
-    benchmark(core.to_newick_from_matrix, core.sample_matrix(sample_size, False))
 
 
 @pytest.mark.parametrize("sample_size", BIG_RANGE)
@@ -81,3 +67,17 @@ def test_matrix_to_newick(benchmark, sample_size):
         core.to_matrix,
         core.to_newick_from_matrix(core.sample_matrix(sample_size, False)),
     )
+
+
+@pytest.mark.parametrize("sample_size", BIG_RANGE)
+def test_matrix_from_newick(benchmark, sample_size):
+    """Benchmark the conversion of an unordered matrix to Newick format.
+
+    Parameters
+    ----------
+    benchmark : _pytest.fixtures.BenchmarkFixture
+        Benchmark fixture from pytest-benchmark
+    sample_size : int
+        Number of leaves in the vector
+    """
+    benchmark(core.to_newick_from_matrix, core.sample_matrix(sample_size, False))
