### 2025-06-25 addea7e
chore: add benchmarks for base matrix operations (#79)
* feat: add benchmarks for base matrix operations

* chore: add matrix benchmark tests in python

* fix (py): fix benchmark test names
