### 2025-06-26 497ee05
feat: port matrix operations to ndarray + rust-numpy
