### 2025-06-26 aff44fc
Update r-phylo2vec/src/rust/src/lib.rs
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
