### 2025-06-29 69ce418
feat: port matrix operations to ndarray + rust-numpy (#81)
* feat: port matrix operations to ndarray + rust-numpy

* Update r-phylo2vec/src/rust/src/lib.rs

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
