+++ py-phylo2vec/phylo2vec/io/reader.py
@@ -36,7 +36,7 @@ def load(filepath: str, delimiter: str = ",") -> np.ndarray:
     if arr.ndim == 1 and np.issubdtype(arr.dtype, np.integer):
         check_vector(arr)
     elif arr.ndim == 2 and np.issubdtype(arr.dtype, np.floating):
-        check_matrix(arr)
+        check_matrix(arr.astype(np.float32))
     else:
         raise ValueError(
             "Input file should either be a vector (ndim == 1) or matrix (ndim == 2)"
