+++ py-phylo2vec/Cargo.toml
@@ -21,12 +21,15 @@ crate-type = ["cdylib"]
 
 [dependencies]
 rand = "*"
-pyo3 = { version = "0.24.1", features = [
+pyo3 = { version = "0.25.0", features = [
     "abi3-py312",
     "abi3-py311",
     "abi3-py310",
+    "extension-module"
 ] }
 phylo2vec = { workspace = true }
+numpy = "0.25.0"
+ndarray = "0.16.1"
 
 [dev-dependencies]
 rstest = "0.23.0"

[…]
