### 2025-06-30 6e8405e
refactor: re-vamp rust core
