+++ r-phylo2vec/src/rust/src/lib.rs
@@ -1,10 +1,16 @@
 use extendr_api::prelude::*;
-use ndarray::{Array2, Axis};
+use ndarray::Array2;
 use std::collections::HashMap;
 use std::result::Result;
 
-use phylo2vec::tree_vec::ops;
-use phylo2vec::utils;
+use phylo2vec::matrix::base as mbase;
+use phylo2vec::matrix::convert as mconvert;
+use phylo2vec::matrix::graph as mgraph;
+use phylo2vec::newick;
+use phylo2vec::vector::base as vbase;
+use phylo2vec::vector::convert as vconvert;
+use phylo2vec::vector::graph as vgraph;
+use phylo2vec::vector::ops as vops;
 
 fn as_usize(v: Vec<i32>) -> Vec<usize> {
     v.iter().map(|&x| x as usize).collect()
@@ -53,7 +59,7 @@ fn convert_from_rmatrix(matrix: &Robj) -> Result<Array2<f32>, &'static str> {
 /// @export
 #[extendr]
 fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
-    let v = utils::sample_vector(n_leaves, ordered);
+    let v = vbase::sample_vector(n_leaves, ordered);
     as_i32(v)
 }
 
@@ -61,7 +67,7 @@ fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
 /// @export
 #[extendr]
 fn sample_matrix(n_leaves: usize, ordered: bool) -> RMatrix<f64> {
-    let matrix = utils::sample_matrix(n_leaves, ordered);
+    let matrix = mbase::sample_matrix(n_leaves, ordered);
     let shape = matrix.shape();
     RMatrix::new_matrix(shape[0], shape[1], |r, c| matrix[[r, c]] as f64)
 }
@@ -71,7 +77,7 @@ fn sample_matrix(n_leaves: usize, ordered: bool) -> RMatrix<f64> {
 #[extendr]
 fn to_newick_from_vector(vector: Vec<i32>) -> String {
     let v_usize = as_usize(vector);
-    ops::to_newick_from_vector(&v_usize)
+    vconvert::to_newick(&v_usize)
 }
 
 /// Recover a rooted tree (in Newick format) from a Phylo2Vec matrix
@@ -79,14 +85,14 @@ fn to_newick_from_vector(vector: Vec<i32>) -> String {
 #[extendr]
 fn to_newick_from_matrix(matrix: RMatrix<f64>) -> String {
     let matrix = convert_from_rmatrix(&matrix).unwrap();
-    ops::to_newick_from_matrix(&matrix.view())
+    mconvert::to_newick(&matrix.view())
 }
 
 /// Convert a newick string to a Phylo2Vec vector
 /// @export
 #[extendr]
 fn to_vector(newick: &str) -> Vec<i32> {
-    let v = ops::to_vector(newick);
+    let v = vconvert::from_newick(newick);
     as_i32(v)
 }
 
@@ -94,7 +100,7 @@ fn to_vector(newick: &str) -> Vec<i32> {
 /// @export
 #[extendr]
 fn to_matrix(newick: &str) -> RMatrix<f64> {
-    let matrix = ops::matrix::to_matrix(newick);
+    let matrix = mconvert::from_newick(newick);
     let shape = matrix.shape();
     RMatrix::new_matrix(shape[0], shape[1], |r, c| matrix[[r, c]] as f64)
 }
@@ -104,7 +110,7 @@ fn to_matrix(newick: &str) -> RMatrix<f64> {
 #[extendr]
 fn check_v(vector: Vec<i32>) {
     let v_usize = as_usize(vector);
-    utils::check_v(&v_usize);
[…]
