### 2025-06-30 6dc3ed7
Revamp rust core and port matrix operations to ndarray (#81) (#82)
* feat: port matrix operations to ndarray + rust-numpy

* refactor: re-vamp rust core

* Trigger Pre-Commit CI

* fix(demo): fix matrix dtype to float32
