+++ r-phylo2vec/src/rust/src/lib.rs
@@ -1,9 +1,16 @@
 use extendr_api::prelude::*;
+use ndarray::Array2;
 use std::collections::HashMap;
 use std::result::Result;
 
-use phylo2vec::tree_vec::ops;
-use phylo2vec::utils;
+use phylo2vec::matrix::base as mbase;
+use phylo2vec::matrix::convert as mconvert;
+use phylo2vec::matrix::graph as mgraph;
+use phylo2vec::newick;
+use phylo2vec::vector::base as vbase;
+use phylo2vec::vector::convert as vconvert;
+use phylo2vec::vector::graph as vgraph;
+use phylo2vec::vector::ops as vops;
 
 fn as_usize(v: Vec<i32>) -> Vec<usize> {
     v.iter().map(|&x| x as usize).collect()
@@ -14,27 +21,45 @@ fn as_i32(v: Vec<usize>) -> Vec<i32> {
 }
 
 // Convert R matrix to Rust Vec<Vec<f32>>
-fn convert_from_rmatrix(matrix: &Robj) -> Result<Vec<Vec<f32>>, &'static str> {
+fn convert_from_rmatrix(matrix: &Robj) -> Result<Array2<f32>, &'static str> {
     let data = matrix.as_real_slice().ok_or("Expected numeric matrix")?;
     let dims = matrix.dim().ok_or("Matrix is missing dimensions")?;
 
     let (nrows, ncols) = (dims[0].inner() as usize, dims[1].inner() as usize);
 
+    if data.len() != nrows * ncols {
+        return Err("Matrix dimensions do not match data length");
+    }
+    // Create a 2D array with the specified dimensions
+    let mut array = Array2::<f32>::zeros((nrows, ncols));
+    // for (i, mut row) in array.axis_iter_mut(Axis(0)).enumerate() {
+    //     for j in 0..ncols {
+    //         row[j] = data[] as f32; // Convert to f32
+    //     }
+    // }
+    for i in 0..nrows {
+        for j in 0..ncols {
+            array[[i, j]] = data[j * nrows + i] as f32; // Convert to f32
+        }
+    }
+
+    Ok(array)
+
     // Convert column-major to row-major Vec<Vec<f32>>
-    Ok((0..nrows)
-        .map(|row| {
-            (0..ncols)
-                .map(|col| data[col * nrows + row] as f32)
-                .collect()
-        })
-        .collect())
+    // Ok((0..nrows)
+    //     .map(|row| {
+    //         (0..ncols)
+    //             .map(|col| data[col * nrows + row] as f32)
+    //             .collect()
+    //     })
+    //     .collect())
 }
 
 /// Sample a random tree topology via Phylo2Vec
 /// @export
 #[extendr]
 fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
-    let v = utils::sample_vector(n_leaves, ordered);
+    let v = vbase::sample_vector(n_leaves, ordered);
     as_i32(v)
 }
 
@@ -42,10 +67,9 @@ fn sample_vector(n_leaves: usize, ordered: bool) -> Vec<i32> {
 /// @export
 #[extendr]
 fn sample_matrix(n_leaves: usize, ordered: bool) -> RMatrix<f64> {
-    let matrix = utils::sample_matrix(n_leaves, ordered);
-    let nrows = matrix.len();
-    let ncols = matrix[0].len();
-    RMatrix::new_matrix(nrows, ncols, |r, c| matrix[r][c] as f64)
[…]
