+++ py-phylo2vec/src/lib.rs
@@ -1,138 +1,152 @@
+use std::collections::HashMap;
+
+use numpy::{IntoPyArray, PyArray2, PyReadonlyArray2};
 use pyo3::prelude::*;
 
-use phylo2vec::tree_vec::ops;
-use phylo2vec::utils;
-use std::collections::HashMap;
+use phylo2vec::matrix::base as mbase;
+use phylo2vec::matrix::convert as mconvert;
+use phylo2vec::matrix::graph as mgraph;
+use phylo2vec::newick;
+use phylo2vec::vector::base as vbase;
+use phylo2vec::vector::convert as vconvert;
+use phylo2vec::vector::graph as vgraph;
+use phylo2vec::vector::ops as vops;
 
 #[pyfunction]
 fn to_newick_from_vector(input_vector: Vec<usize>) -> PyResult<String> {
-    let newick = ops::to_newick_from_vector(&input_vector);
+    let newick = vconvert::to_newick(&input_vector);
     Ok(newick)
 }
 
 #[pyfunction]
-fn to_newick_from_matrix(input_matrix: Vec<Vec<f32>>) -> PyResult<String> {
-    let newick = ops::to_newick_from_matrix(&input_matrix);
+fn to_newick_from_matrix(input_matrix: PyReadonlyArray2<f32>) -> PyResult<String> {
+    let arr = input_matrix.as_array();
+    let newick = mconvert::to_newick(&arr);
     Ok(newick)
 }
 
 #[pyfunction]
 fn to_vector(newick: &str) -> Vec<usize> {
-    ops::to_vector(newick)
+    vconvert::from_newick(newick)
 }
 
 #[pyfunction]
-fn to_matrix(newick: &str) -> Vec<Vec<f32>> {
-    ops::matrix::to_matrix(newick)
+fn to_matrix<'py>(py: Python<'py>, newick: &str) -> Bound<'py, PyArray2<f32>> {
+    mconvert::from_newick(newick).into_pyarray(py)
 }
 
 #[pyfunction]
 fn get_ancestry(input_vector: Vec<usize>) -> Vec<[usize; 3]> {
-    let ancestry: Vec<[usize; 3]> = ops::get_ancestry(&input_vector);
+    let ancestry: Vec<[usize; 3]> = vconvert::to_ancestry(&input_vector);
 
     ancestry
 }
 
 #[pyfunction]
 fn from_ancestry(input_ancestry: Vec<[usize; 3]>) -> Vec<usize> {
-    let vector: Vec<usize> = ops::from_ancestry(&input_ancestry);
+    let vector: Vec<usize> = vconvert::from_ancestry(&input_ancestry);
     vector
 }
 
 #[pyfunction]
 fn get_pairs(input_vector: Vec<usize>) -> Vec<(usize, usize)> {
-    ops::get_pairs(&input_vector)
+    vconvert::to_pairs(&input_vector)
 }
 
 #[pyfunction]
 fn from_pairs(input_pairs: Vec<(usize, usize)>) -> Vec<usize> {
-    ops::from_pairs(&input_pairs)
+    vconvert::from_pairs(&input_pairs)
 }
 
 #[pyfunction]
 fn get_edges(input_vector: Vec<usize>) -> Vec<(usize, usize)> {
-    ops::get_edges(&input_vector)
+    vconvert::to_edges(&input_vector)
 }
 
 #[pyfunction]
 fn from_edges(input_edges: Vec<(usize, usize)>) -> Vec<usize> {
-    ops::from_edges(&input_edges)
+    vconvert::from_edges(&input_edges)
 }
 
 #[pyfunction]
 fn build_newick(input_pairs: Vec<(usize, usize)>) -> String {
-    let newick_string: String = ops::newick::build_newick(&input_pairs);
[…]
