### 2025-07-02 257e8c0
feat: vcov and precision matrices + move float ops to f64
