+++ py-phylo2vec/phylo2vec/stats/__init__.py
@@ -1,0 +1,7 @@
+"""
+Methods to convert Phylo2Mat matrices to Newick format and vice-versa.
+"""
+
+from .nodewise import cophenetic_distances, cov, pairwise_distances, precision
+
+__all__ = ["cophenetic_distances", "cov", "pairwise_distances", "precision"]

[…]
