### 2025-07-02 c1b8015
docs(r): fix docs for new functions in R
