+++ r-phylo2vec/R/stats.R
@@ -1,11 +1,8 @@
-#' Compute the cophenetic distance matrix of a precursorhylo2Vec
+#' Compute the cophenetic distance matrix of a Phylo2Vec
 #' vector (topological) or matrix (from branch lengths).
 #'
-#' @param n_leaves Number of leaves in the tree
-#' @param topology_only If TRUE, a vector corresponding to the topology
-#' is returned. IF FALSE, a matrix is returned
-#' @param ordered If TRUE, sample an ordered tree
-#' @return A precursorhylo2Vec object (`integer` if topology_only, else `matrix`)
+#' @param vector_or_matrix Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)
+#' @return Cophenetic distance matrix
 #' @export
 cophenetic_distances <- function(vector_or_matrix) {
   if (is.vector(vector_or_matrix)) {
@@ -19,6 +16,12 @@ cophenetic_distances <- function(vector_or_matrix) {
   }
 }
 
+#' Compute the variance-covariance matrix of a Phylo2Vec
+#' vector (topological) or matrix (from branch lengths).
+#'
+#' @param vector_or_matrix Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)
+#' @return Variance-covariance matrix
+#' @export
 cov <- function(vector_or_matrix) {
   if (is.vector(vector_or_matrix)) {
     # If the input is a vector, call the C function for vector
@@ -31,6 +34,14 @@ cov <- function(vector_or_matrix) {
   }
 }
 
+#' Compute the precision matrix of a Phylo2Vec
+#' vector (topological) or matrix (from branch lengths).
+#'
+#' The precision matrix is the inverse of the variance-covariance matrix.
+#'
+#' @param vector_or_matrix Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)
+#' @return Precision matrix
+#' @export
 precision <- function(vector_or_matrix) {
   if (is.vector(vector_or_matrix)) {
     # If the input is a vector, call the C function for vector
@@ -42,6 +53,7 @@ precision <- function(vector_or_matrix) {
     stop("Input must be either a vector or a 2D matrix.")
   }
 
+  # Schur complement of the precursor matrix
   # nrow(precursor) = ncol(precursor) = 2*k where k = n_leaves - 1
   n <- nrow(precursor)
   n_leaves <- nrow(precursor) / 2 + 1
