+++ r-phylo2vec/man/cov.Rd
@@ -1,19 +1,19 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/stats.R
-\name{cophenetic_distances}
-\alias{cophenetic_distances}
-\title{Compute the cophenetic distance matrix of a Phylo2Vec
+\name{cov}
+\alias{cov}
+\title{Compute the variance-covariance matrix of a Phylo2Vec
 vector (topological) or matrix (from branch lengths).}
 \usage{
-cophenetic_distances(vector_or_matrix)
+cov(vector_or_matrix)
 }
 \arguments{
 \item{vector_or_matrix}{Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)}
 }
 \value{
-Cophenetic distance matrix
+Variance-covariance matrix
 }
 \description{
-Compute the cophenetic distance matrix of a Phylo2Vec
+Compute the variance-covariance matrix of a Phylo2Vec
 vector (topological) or matrix (from branch lengths).
 }
