+++ r-phylo2vec/man/precision.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/stats.R
+\name{precision}
+\alias{precision}
+\title{Compute the precision matrix of a Phylo2Vec
+vector (topological) or matrix (from branch lengths).}
+\usage{
+precision(vector_or_matrix)
+}
+\arguments{
+\item{vector_or_matrix}{Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)}
+}
+\value{
+Precision matrix
+}
+\description{
+The precision matrix is the inverse of the variance-covariance matrix.
+}
