+++ r-phylo2vec/NAMESPACE
@@ -7,6 +7,7 @@ export(check_v)
 export(cophenetic_distances)
 export(cophenetic_from_matrix)
 export(cophenetic_from_vector)
+export(cov)
 export(create_label_mapping)
 export(from_ancestry)
 export(from_edges)
@@ -18,6 +19,7 @@ export(load_newick)
 export(load_p2v)
 export(pre_precision_from_matrix)
 export(pre_precision_from_vector)
+export(precision)
 export(queue_shuffle)
 export(remove_leaf)
 export(sample_matrix)
