### 2025-07-02 9746236
refactor(r/lib): remove deprecated f32 comment
