### 2025-07-02 3c0e7db
docs(py): fix docs in stats
