### 2025-07-02 3c57303
Update phylo2vec/src/matrix/graph.rs
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
