### 2025-07-02 0fb7b98
style: pre-commit fixes
