### 2025-07-02 de55cda
refactor: rename variance-covariance functions
