+++ r-phylo2vec/man/vcov_from_vector.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/extendr-wrappers.R
-\name{vcv_from_vector}
-\alias{vcv_from_vector}
+\name{vcov_from_vector}
+\alias{vcov_from_vector}
 \title{Get the variance-covariance matrix of a Phylo2Vec vector}
 \usage{
-vcv_from_vector(vector)
+vcov_from_vector(vector)
 }
 \description{
 Get the variance-covariance matrix of a Phylo2Vec vector
