+++ r-phylo2vec/man/vcov_from_matrix.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/extendr-wrappers.R
-\name{vcv_from_matrix}
-\alias{vcv_from_matrix}
+\name{vcov_from_matrix}
+\alias{vcov_from_matrix}
 \title{Get the variance-covariance matrix of a Phylo2Vec matrix}
 \usage{
-vcv_from_matrix(matrix)
+vcov_from_matrix(matrix)
 }
 \description{
 Get the variance-covariance matrix of a Phylo2Vec matrix
