+++ r-phylo2vec/NAMESPACE
@@ -7,7 +7,6 @@ export(check_v)
 export(cophenetic_distances)
 export(cophenetic_from_matrix)
 export(cophenetic_from_vector)
-export(cov)
 export(create_label_mapping)
 export(from_ancestry)
 export(from_edges)
@@ -34,6 +33,7 @@ export(to_newick_from_matrix)
 export(to_newick_from_vector)
 export(to_pairs)
 export(to_vector)
-export(vcv_from_matrix)
-export(vcv_from_vector)
+export(vcov_from_matrix)
+export(vcov_from_vector)
+export(vcovp)
 useDynLib(phylo2vec, .registration = TRUE)
