### 2025-07-02 b15a67b
fix(r): fix vcv name in stats
