### 2025-07-02 f08fad4
feat: vcov and precision matrices + move float ops to f64 (#86)
* feat: vcov and precision matrices + move float ops to f64

* style: pre-commit fixes

* docs(r): fix docs for new functions in R

* refactor(py): add legacy alias for to-be-deprecated metrics

* refactor(r/lib): remove deprecated f32 comment

* docs(py): fix docs in stats

* Update phylo2vec/src/matrix/graph.rs

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* style: pre-commit fixes

* refactor: rename variance-covariance functions

* fix(r): fix vcv name in stats

---------

Co-authored-by: pre-commit-ci
[…]
