+++ py-phylo2vec/phylo2vec/metrics/__init__.py
@@ -1,7 +1,0 @@
-"""
-Methods to convert Phylo2Mat matrices to Newick format and vice-versa.
-"""
-
-from .pairwise import cophenetic_distances, pairwise_distances
-
-__all__ = ["cophenetic_distances", "pairwise_distances"]
