### 2025-07-03 d5a83aa
feat: add incidence matrix functions
