+++ r-phylo2vec/R/extendr-wrappers.R
@@ -73,6 +73,18 @@ get_common_ancestor <- function(vector, node1, node2) .Call(wrap__get_common_anc
 #' @export
 has_branch_lengths <- function(newick) .Call(wrap__has_branch_lengths, newick)
 
+#' Get the incidence matrix of a Phylo2Vec vector in COO format
+#' @export
+incidence_coo <- function(input_vector) .Call(wrap__incidence_coo, input_vector)
+
+#' Get the incidence matrix of a Phylo2Vec vector in CSR format
+#' @export
+incidence_csr <- function(input_vector) .Call(wrap__incidence_csr, input_vector)
+
+#' Get the incidence matrix of a Phylo2Vec vector in dense format
+#' @export
+incidence_dense <- function(input_vector) .Call(wrap__incidence_dense, input_vector)
+
 #' Get the precision matrix of a Phylo2Vec matrix
 #' @export
 pre_precision_from_matrix <- function(matrix) .Call(wrap__pre_precision_from_matrix, matrix)
