+++ phylo2vec/src/vector/graph.rs
@@ -178,20 +178,6 @@ fn get_descendants_and_edges(pairs: &Pairs) -> (HashMap<usize, BitSet>, Vec<(usi
 /// Output is a matrix of dimensions n x n
 /// where n = number of leaves.
 /// Adapted from `vcv.phylo` function in the `ape` package: <https://github.com/emmanuelparadis/ape>
-/// Example:
-/// ```
-/// use ndarray::array;
-/// use phylo2vec::vector::graph::vcv;
-/// let v = vec![0, 1, 2];
-/// let vcv_matrix = vcv(&v);
-/// let arr = array![
-///     [1.0, 0.0, 0.0, 0.0],
-///     [0.0, 2.0, 1.0, 1.0],
-///     [0.0, 1.0, 3.0, 2.0],
-///     [0.0, 1.0, 2.0, 3.0]
-/// ];
-/// assert_eq!(vcv_matrix, arr);
-/// ```
 pub fn _vcv(v: &[usize], bls: Option<&Vec<[f64; 2]>>) -> Array2<f64> {
     let k = v.len();
     let n_leaves = k + 1;
@@ -235,8 +221,243 @@ pub fn _vcv(v: &[usize], bls: Option<&Vec<[f64; 2]>>) -> Array2<f64> {
     out
 }
 
+/// Get variance-covariance matrix
+/// for a Phylo2Vec vector.
+/// Output is a matrix of dimensions n x n
+/// where n = number of leaves.
+/// Example:
+/// ```
+/// use ndarray::array;
+/// use phylo2vec::vector::graph::vcv;
+/// let v = vec![0, 1, 2];
+/// let vcv_matrix = vcv(&v);
+/// let arr = array![
+///     [1.0, 0.0, 0.0, 0.0],
+///     [0.0, 2.0, 1.0, 1.0],
+///     [0.0, 1.0, 3.0, 2.0],
+///     [0.0, 1.0, 2.0, 3.0]
+/// ];
+/// assert_eq!(vcv_matrix, arr);
+/// ```
 pub fn vcv(v: &[usize]) -> Array2<f64> {
     _vcv(v, None)
+}
+
+/// The `Incidence` struct represents an incidence matrix for a phylogenetic tree
+/// derived from a Phylo2Vec vector.
+/// Note that the Phylo2Vec vector represents a rooted binary tree, which can be
+/// viewed as a directed acyclic graph (DAG). Duplicates are not allowed.
+///
+/// # Fields
+///
+/// - `n_nodes`: The total number of nodes in the tree. This is calculated as `2*k + 1`,
+///   where `k` is the length of the input Phylo2Vec vector.
+/// - `n_edges`: The total number of edges in the tree. This is calculated as `2*k`,
+///   where `k` is the length of the input Phylo2Vec vector.
+/// - `edges`: A vector of tuples where each tuple `(usize, usize)` represents a directed
+///   edge in the tree. In each tuple, the first element is the child node and the second
+///   element is the parent node.
+///
+/// # Methods
+///
+/// - `new(v: &[usize]) -> Self`
+///   - Constructs a new `Incidence` instance from a Phylo2Vec vector `v`.
+/// - `to_dense(&self) -> Array2<i8>`
+///   - Returns the incidence matrix in a dense format as a 2D ndarray of type `i8`.
+/// - `to_coo(&self) -> (Vec<i8>, Vec<usize>, Vec<usize>)`
[…]
