+++ r-phylo2vec/NAMESPACE
@@ -14,6 +14,9 @@ export(from_newick)
 export(from_pairs)
 export(get_common_ancestor)
 export(has_branch_lengths)
+export(incidence_coo)
+export(incidence_csr)
+export(incidence_dense)
 export(load_newick)
 export(load_p2v)
 export(pre_precision_from_matrix)

[…]
