### 2025-07-03 47c3f13
style: pre-commit fixes
