### 2025-07-04 8fe1f8d
feat(r): add incidence function and tests
