+++ r-phylo2vec/src/rust/src/lib.rs
@@ -392,11 +392,13 @@ fn vcov_from_matrix(matrix: RMatrix<f64>) -> RMatrix<f64> {
 #[extendr]
 fn incidence_dense(input_vector: Vec<i32>) -> RMatrix<i32> {
     let v_usize = as_usize(input_vector);
+    let k = v_usize.len();
     let dense_rs = vgraph::Incidence::new(&v_usize).to_dense();
-    let n_leaves = dense_rs.shape()[0];
-    let mut dense_r = RMatrix::new_matrix(n_leaves, n_leaves, |r, c| dense_rs[[r, c]] as i32);
-    let dimnames = (0..n_leaves).map(|x| x as i32).collect::<Vec<i32>>();
-    dense_r.set_dimnames(List::from_values(vec![dimnames.clone(), dimnames]));
+    let mut dense_r = RMatrix::new_matrix(2 * k + 1, 2 * k, |r, c| dense_rs[[r, c]] as i32);
+    let colnames = (0..2 * k).map(|x| x as i32).collect::<Vec<i32>>();
+    let mut rownames = colnames.clone();
+    rownames.push(2 * k as i32);
+    dense_r.set_dimnames(List::from_values(vec![rownames, colnames]));
 
     dense_r
 }
@@ -404,16 +406,27 @@ fn incidence_dense(input_vector: Vec<i32>) -> RMatrix<i32> {
 /// Get the incidence matrix of a Phylo2Vec vector in COO format
 /// @export
 #[extendr]
-fn incidence_coo(input_vector: Vec<i32>) -> RMatrix<i32> {
-    let k = input_vector.len();
+fn incidence_coo(input_vector: Vec<i32>) -> List {
     let v_usize = as_usize(input_vector);
     let (data, rows, cols) = vgraph::Incidence::new(&v_usize).to_coo();
-    RMatrix::new_matrix(4 * k, 3, |r, c| match c {
-        0 => data[r] as i32,
-        1 => rows[r] as i32,
-        2 => cols[r] as i32,
-        _ => unreachable!(),
-    })
+    list!(
+        data = data.iter().map(|&x| x as i32).collect::<Vec<i32>>(),
+        rows = as_i32(rows),
+        cols = as_i32(cols)
+    )
+}
+
+/// Get the incidence matrix of a Phylo2Vec vector in CSR format
+/// @export
+#[extendr]
+fn incidence_csc(input_vector: Vec<i32>) -> List {
+    let v_usize = as_usize(input_vector);
+    let (data, indices, indptr) = vgraph::Incidence::new(&v_usize).to_csc();
+    list!(
+        data = data.iter().map(|&x| x as i32).collect::<Vec<i32>>(),
+        indices = as_i32(indices),
+        indptr = as_i32(indptr)
+    )
 }
 
 /// Get the incidence matrix of a Phylo2Vec vector in CSR format
@@ -447,6 +460,7 @@ extendr_module! {
     fn get_common_ancestor;
     fn has_branch_lengths;
     fn incidence_coo;
+    fn incidence_csc;
     fn incidence_csr;
     fn incidence_dense;
     fn pre_precision_from_matrix;
