+++ r-phylo2vec/NAMESPACE
@@ -14,7 +14,9 @@ export(from_newick)
 export(from_pairs)
 export(get_common_ancestor)
 export(has_branch_lengths)
+export(incidence)
 export(incidence_coo)
+export(incidence_csc)
 export(incidence_csr)
 export(incidence_dense)
 export(load_newick)
