### 2025-07-04 d90240b
refactor: discard precommit codespell
