+++ r-phylo2vec/src/rust/src/lib.rs
@@ -403,7 +403,7 @@ fn incidence_dense(input_vector: Vec<i32>) -> RMatrix<i32> {
     dense_r
 }
 
-/// Get the incidence matrix of a Phylo2Vec vector in COUP format
+/// Get the incidence matrix of a Phylo2Vec vector in COO format
 /// @export
 #[extendr]
 fn incidence_coo(input_vector: Vec<i32>) -> List {
