### 2025-07-04 bd3c3da
chore: add scipy to pixi dependencies for testing
