### 2025-07-04 575708b
refactor(py/incidence): remove dok, make coo default and add tests
