### 2025-07-04 0a7b3e8
refactor(rust): discard precommit codespell
