### 2025-07-04 91f2492
feat: add incidence matrix functions (#89)
* feat: add incidence matrix functions

* style: pre-commit fixes

* feat(r): add incidence function and tests

* refactor: add coo/COO to codespell whitelist

* refactor: discard precommit codespell

* chore: add scipy to pixi dependencies for testing

* refactor(py/incidence): remove dok, make coo default and add tests

* refactor(rust): discard precommit codespell

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
