+++ r-phylo2vec/man/incidence_csc.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{incidence_csc}
+\alias{incidence_csc}
+\title{Get the incidence matrix of a Phylo2Vec vector in CSR format}
+\usage{
+incidence_csc(input_vector)
+}
+\description{
+Get the incidence matrix of a Phylo2Vec vector in CSR format
+}

[…]
