### 2025-07-04 51420fd
refactor(py): code cleaning
