+++ py-phylo2vec/phylo2vec/io/_validation.py
@@ -2,7 +2,6 @@
 
 from pathlib import Path
 
-
 FILE_EXTENSIONS = {
     "array": [".csv", ".txt"],
     "newick": [".txt", ".nwk", ".newick", ".tree", ".treefile"],
@@ -12,7 +11,8 @@ FILE_EXTENSIONS = {
 def check_path(filepath: str, file_type: str):
     suffix = Path(filepath).suffix
     file_extensions = FILE_EXTENSIONS[file_type]
-    assert suffix in file_extensions, (
-        f"Unsupported file extension: {suffix}. "
-        f"Accepted extensions for {file_type} files: {file_extensions}"
-    )
+    if suffix not in file_extensions:
+        raise ValueError(
+            f"Unsupported file extension: {suffix}. "
+            f"Accepted extensions for {file_type} files: {file_extensions}"
+        )
