+++ resources/update_version.py
@@ -1,8 +1,10 @@
-import sys
-import os
-import subprocess
-import re
 import argparse
+import os
+import re
+import shutil
+import subprocess
+import sys
+
 from pathlib import Path
 from typing import Tuple
 
@@ -16,6 +18,7 @@ except ImportError:
 
 
 # ---- Common utility functions ----
+
 
 def increment_patch_version(version_str: str) -> str:
     """
@@ -32,16 +35,18 @@ def increment_patch_version(version_str: str) -> str:
         Version with incremented patch number
     """
     # Extract the base version (remove any suffixes like -rc.xxx or .devxxx)
-    base_version = re.match(r'(\d+\.\d+\.\d+)', version_str).group(1)
+    base_version = re.match(r"(\d+\.\d+\.\d+)", version_str).group(1)
 
     if not base_version:
         print(f"Warning: Could not parse version string '{version_str}'")
         return version_str
 
     # Split into major, minor, patch
-    parts = base_version.split('.')
+    parts = base_version.split(".")
     if len(parts) != 3:
-        print(f"Warning: Version '{base_version}' doesn't follow semver major.minor.patch format")
+        print(
+            f"Warning: Version '{base_version}' doesn't follow semver major.minor.patch format"
+        )
         return version_str
 
     # Increment patch version
@@ -69,23 +74,35 @@ def get_distance_from_version_tag(version_str):
         The distance from the tag, or an empty string if git command fails
     """
     try:
+        git_executable = shutil.which("git")
+        if not git_executable:
+            raise RuntimeError("`git` command not found. Please install Git.")
+
         # Clean version string for tag matching (remove any suffixes)
-        base_version = re.match(r'(\d+\.\d+\.\d+)', version_str).group(1)
+        base_version = re.match(r"(\d+\.\d+\.\d+)", version_str).group(1)
 
         # Format for tag search
         tag_prefix = f"v{base_version}"
 
         # First check if the tag exists
-        result = subprocess.run(['git', 'tag', '-l', tag_prefix],  # skipcq: BAN-B607
-                               capture_output=True, text=True, check=True)
+        result = subprocess.run(
+            [git_executable, "tag", "-l", tag_prefix],  # skipcq: BAN-B607
+            capture_output=True,
+            text=True,
+            check=True,
+        )
 
         if not result.stdout.strip():
             print(f"Warning: No tag found matching '{tag_prefix}'")
             return "0"  # Return 0 if no tag found
 
         # Get the distance from tag to HEAD
-        result = subprocess.run(['git', 'rev-list', f'{tag_prefix}..HEAD', '--count'],  # skipcq: BAN-B607
-                               capture_output=True, text=True, check=True)
+        result = subprocess.run(
+            [git_executable, "rev-list", f"{tag_prefix}..HEAD", "--count"],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
         distance = result.stdout.strip()
         return distance
     except (subprocess.SubprocessError, FileNotFoundError):
@@ -94,6 +111,7 @@ def get_distance_from_version_tag(version_str):
 
 
 # ---- Python/Cargo.toml handling ----
+
 
 def get_python_cargo_path() -> Path:
     """Get the path to the Python binding's Cargo.toml file."""
@@ -128,12 +146,13 @@ def get_python_version(cargo_path: Path) -> str:
 
     try:
         toml_dict = tomllib.loads(content)
-        if 'package' in toml_dict and 'version' in toml_dict['package']:
-            return toml_dict['package']['version']
-        else:
-            raise ValueError("No version field found in the package section of the TOML file")
-    except Exception as e:
-        raise ValueError(f"Failed to parse TOML file: {e}")
+        if "package" in toml_dict and "version" in toml_dict["package"]:
[…]
