### 2025-07-04 7f5253e
refactor(rust): add more docstrings to public functions
