+++ phylo2vec/src/vector/convert.rs
@@ -6,6 +6,20 @@ use crate::types::{Ancestry, Pair, Pairs};
 use crate::vector::avl::AVLTree;
 use crate::vector::base::is_unordered;
 
+/// Retrieve a Phylo2Vec vector from a list of pairs.
+/// A pair is a tuple of nodes (c1, c2) where:
+///   - c1 denotes the node from where the branch leading to c2 originates
+///   - c2 is the leaf node
+///
+/// # Examples
+///
+/// ```
+/// use phylo2vec::vector::convert::from_pairs;
+///
+/// let pairs = vec![(1, 4), (0, 3), (0, 2), (0, 1)];
+/// let v = from_pairs(&pairs);
+/// assert_eq!(v, vec![0, 0, 0, 1]);
+/// ```
 pub fn from_pairs(pairs: &Pairs) -> Vec<usize> {
     let mut cherries: Ancestry = Vec::with_capacity(pairs.len());
     for &(c1, c2) in pairs.iter() {
@@ -77,12 +91,14 @@ fn to_pairs_avl(v: &[usize]) -> Pairs {
 /// AVL trees are faster for unordered trees.
 /// A simple for loop is faster for ordered trees.
 ///
-/// # Example
+/// # Examples
+///
 /// ```
 /// use phylo2vec::vector::convert::to_pairs;
 ///
-/// let v = vec![0, 0, 0, 1, 3, 3, 1, 4, 4];
+/// let v = vec![0, 0, 0, 1];
 /// let pairs = to_pairs(&v);
+/// assert_eq!(pairs, vec![(1, 4), (0, 3), (0, 2), (0, 1)]);
 /// ```
 pub fn to_pairs(v: &[usize]) -> Pairs {
     if is_unordered(v) {
@@ -92,6 +108,24 @@ pub fn to_pairs(v: &[usize]) -> Pairs {
     }
 }
 
+/// Retrieve a Phylo2Vec vector from an ancestry representation.
+/// The ancestry is a list of cherries, where each cherry is represented as
+/// a triplet [c1, c2, p], where:
+///     - c1 and c2 are the indices of the leaves in the cherry
+///     - p is the index of the parent node of the cherry
+///
+/// # Examples
+///
+/// ```
+/// use phylo2vec::vector::convert::from_ancestry;
+/// let ancestry = vec![
+///    [2, 3, 4], // Cherry with leaves 0 and 1, parent 4
+///    [0, 1, 5], // Cherry with leaves 0 and 2, parent 5
+///    [5, 4, 6], // Cherry with nodes 4 and 5, parent 6
+/// ];
+/// let v = from_ancestry(&ancestry);
+/// assert_eq!(v, vec![0, 2, 2]);
+/// ```
 pub fn from_ancestry(ancestry: &Ancestry) -> Vec<usize> {
     let mut cherries = ancestry.clone();
 
@@ -102,7 +136,7 @@ pub fn from_ancestry(ancestry: &Ancestry) -> Vec<usize> {
 }
 
 /// Get the ancestry of the Phylo2Vec vector
-/// v[i] = which BRANCH we do the pairing from
+/// v\[i\] = which BRANCH we do the pairing from
 ///
 /// The initial situation looks like this:
 ///                  R
@@ -112,11 +146,24 @@ pub fn from_ancestry(ancestry: &Ancestry) -> Vec<usize> {
 ///  branch 0 <-- //   \\  --> branch 1
 ///               0     1
 ///
-/// For v[1], we have 3 possible branches too choose from.
-/// v[1] = 0 or 1 indicates that we branch out from branch 0 or 1, respectively.
+/// For v\[1\], we have 3 possible branches too choose from.
[…]
