### 2025-07-04 ae98ac0
fix(py): handle the no-match case gracefully in update_version
