### 2025-07-04 a0c4b8b
refactor(py): print real warnings for update_version and print runtimerror for git
