### 2025-07-04 840c625
Cleaning (#90)
* refactor(py): code cleaning

* refactor(rust): add more docstrings to public functions

* style: pre-commit fixes

* fix(py): fix error catching for io and simplify functions

* Update phylo2vec/src/vector/convert.rs

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* fix(py): handle the no-match case gracefully in update_version

* refactor(py): print real warnings for update_version and print runtimerror for git

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Copilot <175728472+Copilot
[…]
