+++ .pre-commit-config.yaml
@@ -53,7 +53,7 @@ repos:
         pass_filenames: false
 
   - repo: https://github.com/python-jsonschema/check-jsonschema
-    rev: "0.33.1"
+    rev: "0.33.2"
     hooks:
       - id: check-dependabot
       - id: check-github-workflows
