### 2025-08-04 #8407 @Neclow
Hi @arfon, thank you for the update! I appreciate the information and understand the current backlog situation. Would you be able to provide an estimated time frame for a first decision?

Best regards,

Neil
