### 2025-08-06 d3b17cc
feat: O(n) order_cherries using parental indices
