### 2025-08-07 df2b448
refactor: handle n_leaves edge cases for sampling
