### 2025-08-07 5d73044
refactor: handle n_leaves edge cases for sampling (#95)
