+++ py-phylo2vec/phylo2vec/utils/matrix.py
@@ -24,7 +24,7 @@ def sample_matrix(n_leaves: int, ordered: bool = False) -> np.ndarray:
     Parameters
     ----------
     n_leaves : int
-        Number of leaves
+        Number of leaves (>= 2)
     ordered : bool, optional
         If True, sample an ordered tree, by default False
 
