### 2025-08-07 9cb3e17
feat: add a simple CLI for the python API
