+++ py-phylo2vec/tests/test_cli.py
@@ -1,0 +1,281 @@
+import os
+import sys
+
+from io import StringIO
+
+import numpy as np
+import pytest
+
+from unittest.mock import patch
+
+from phylo2vec.cli import parse_args, main
+
+
+class TestHelp:
+    """Test help functionality."""
+
+    def test_main_help(self):
+        """Test main help message displays correctly."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "--help"]):
+                parse_args()
+        # Help should exit with code 0
+        assert exc_info.value.code == 0
+
+    def test_subcommand_help_samplev(self):
+        """Test samplev subcommand help."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "samplev", "--help"]):
+                parse_args()
+        assert exc_info.value.code == 0
+
+    def test_subcommand_help_samplem(self):
+        """Test samplem subcommand help."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "samplem", "--help"]):
+                parse_args()
+        assert exc_info.value.code == 0
+
+    def test_subcommand_help_from_newick(self):
+        """Test from_newick subcommand help."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "from_newick", "--help"]):
+                parse_args()
+        assert exc_info.value.code == 0
+
+    def test_subcommand_help_to_newick(self):
+        """Test to_newick subcommand help."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "to_newick", "--help"]):
+                parse_args()
+        assert exc_info.value.code == 0
+
+
+class TestSampleV:
+    """Test the samplev command."""
+
+    def test_valid_number(self):
+        """Test samplev with valid number of leaves."""
+        with patch("sys.argv", ["phylo2vec", "samplev", "5"]):
+            args = parse_args()
+            assert hasattr(args, "n_leaves")
+            assert args.n_leaves == 5
+            assert hasattr(args, "ordered")
+            assert not args.ordered
+
+    def test_with_ordered_flag(self):
+        """Test samplev with --ordered flag."""
+        with patch("sys.argv", ["phylo2vec", "samplev", "3", "--ordered"]):
+            args = parse_args()
+            assert args.n_leaves == 3
+            assert args.ordered
+
+    def test_missing_argument(self):
+        """Test samplev without required n_leaves argument."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "samplev"]):
+                parse_args()
+        assert exc_info.value.code != 0
+
+    def test_string(self):
+        """Test samplev with non-numeric string should fail."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.stderr", new_callable=StringIO):  # Suppress error output
+                with patch("sys.argv", ["phylo2vec", "samplev", "not_a_number"]):
+                    parse_args()
+        # Should exit with error code (error 2 for argument errors)
+        assert exc_info.value.code == 2
+
+    def test_float(self):
+        """Test samplev with float number should fail."""
+        with pytest.raises(SystemExit) as exc_info:
+            with patch("sys.argv", ["phylo2vec", "samplev", "3.5"]):
+                parse_args()
+        # Should exit with error code (error 2 for argument errors)
+        assert exc_info.value.code == 2
+
+    def test_negative_number(self):
+        """Test samplev with negative number should fail."""
+        with patch("sys.argv", ["phylo2vec", "samplev", "-5"]):
[…]
