### 2025-08-07 8a247f4
style: pre-commit fixes
