### 2025-08-07 cc26c11
refactor(cli): update pyproject.toml
