### 2025-08-07 e9bb889
refactor(cli): remove unused deps from tests
