### 2025-08-07 0eb5958
feat: re-add 'unrooted' argument to cophenetic distances
