### 2025-08-07 88acaff
feat: re-add 'unrooted' argument to cophenetic distances
