### 2025-08-07 11d4016
feat: add a simple CLI for the python API (#96)
* feat: add a simple CLI for the python API

* style: pre-commit fixes

* refactor(cli): update pyproject.toml

* refactor(cli): remove unused deps from tests

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
