+++ py-phylo2vec/phylo2vec/io/reader.py
@@ -2,6 +2,9 @@
 
 import os
 
+from io import StringIO, TextIOBase
+from typing import Union
+
 import numpy as np
 
 from phylo2vec.base.newick import from_newick
@@ -10,13 +13,19 @@ from phylo2vec.utils.vector import check_vector
 from ._validation import check_path
 
 
-def load(filepath: str, delimiter: str = ",") -> np.ndarray:
+def read(buffer: str, delimiter: str = ",") -> np.ndarray:
+    return load(StringIO(buffer), delimiter=delimiter)
+
+
+def load(
+    filepath_or_buffer: Union[str, TextIOBase], delimiter: str = ","
+) -> np.ndarray:
     """Read a text/csv file into a Phylo2Vec vector or matrix.
 
     Parameters
     ----------
-    filepath : str or file-like object
-        File path to read
+    filepath : str, path object or file-like object
+        File path to read or StringIO object containing the data.
     delimiter : str, optional
         Character used to separate values, by default ","
 
@@ -26,12 +35,15 @@ def load(filepath: str, delimiter: str = ",") -> np.ndarray:
         A vector (ndim == 1) or matrix (ndim == 2)
         which satisfies Phylo2Vec constraints
     """
-    check_path(filepath, "array")
+    if not hasattr(filepath_or_buffer, "read"):
+        if not os.path.isfile(filepath_or_buffer):
+            raise FileNotFoundError(f"File not found: {filepath_or_buffer}")
+        check_path(filepath_or_buffer, "array")
     # np.genfromtxt with dtype = None will infer the type
     # Using should solve the edge case:
     # load a matrix with 2 leaves ([[0, bl1, bl2]] = float)
     # vs. a vector with 4 leaves ([0, v1, v2] = int)
-    arr = np.genfromtxt(filepath, delimiter=delimiter, dtype=None)
+    arr = np.genfromtxt(filepath_or_buffer, delimiter=delimiter, dtype=None)
 
     if arr.ndim == 1 and np.issubdtype(arr.dtype, np.integer):
         check_vector(arr)
