+++ r-phylo2vec/man/cophenetic_distances.Rd
@@ -5,10 +5,13 @@
 \title{Compute the cophenetic distance matrix of a Phylo2Vec
 vector (topological) or matrix (from branch lengths).}
 \usage{
-cophenetic_distances(vector_or_matrix)
+cophenetic_distances(vector_or_matrix, unrooted = FALSE)
 }
 \arguments{
 \item{vector_or_matrix}{Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)}
+
+\item{unrooted}{Logical indicating whether applying the cophenetic distance
+to an unrooted tree or not (see ete3). Default is FALSE.}
 }
 \value{
 Cophenetic distance matrix
