+++ phylo2vec/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "phylo2vec"
 # Rust core version
-version = "0.4.1"
+version = "0.5.0"
 edition.workspace = true
 authors.workspace = true
 description.workspace = true
