+++ r-phylo2vec/src/rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = 'r-phylo2vec'
-version = '1.1.0'
+version = '1.2.0'
 edition.workspace = true
 authors.workspace = true
 description.workspace = true
