+++ .github/workflows/ci-python.yml
@@ -22,8 +22,8 @@ jobs:
       matrix:
         python-environment: ['py310', 'py311', 'py312']
     steps:
-      - uses: actions/checkout@v4
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: actions/checkout@v5
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           environments: ${{ matrix.python-environment }}
       - run: pixi run -e ${{ matrix.python-environment }} test
@@ -31,8 +31,8 @@ jobs:
     name: Python benchmarks
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: actions/checkout@v5
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           environments: py-phylo2vec
       - run: pixi run -e py-phylo2vec benchmark
