+++ .github/workflows/ci-rust.yaml
@@ -23,7 +23,7 @@ jobs:
           # - beta
           # - nightly
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - name: Setup Rust
         run: rustup update ${{ matrix.toolchain }} && rustup default ${{ matrix.toolchain }}
       - name: Run linter
