+++ .github/workflows/ci-R.yml
@@ -19,8 +19,8 @@ jobs:
     name: R project
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: actions/checkout@v5
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           environments: ${{ env.R_ENV }}
       - run: pixi run -e ${{ env.R_ENV }} test "./${{ env.R_ENV}}"
