### 2025-08-18 dc4d75e
chore: update pre-commit hooks
updates:
- [github.com/python-jsonschema/check-jsonschema: 0.33.2 → 0.33.3](https://github.com/python-jsonschema/check-jsonschema/compare/0.33.2...0.33.3)
