### 2025-09-14 #8407 @abhishektiwari
Hi @Neclow I will helping to edit this submission. Can you try to reduce the wordcount for paper to 1000 words?

Also can you confirm if [authorship](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) of the paper is reasonable and appropriate? I can see only 7 contributors to the GitHub repository but 10 authors in paper.
