### 2025-09-15 d9822ee
chore: reduce word count in joss submission (#105)
