+++ joss/paper.bib
@@ -1,15 +1,3 @@
-@article{williams2019,
-  title={Enabling catalyst discovery through machine learning and high-throughput experimentation},
-  author={Williams, Travis and McCullough, Katherine and Lauterbach, Jochen A},
-  journal={Chemistry of Materials},
-  volume={32},
-  number={1},
-  pages={157--165},
-  year={2019},
-  publisher={ACS Publications},
-  doi={10.1021/acs.chemmater.9b03043.s001}
-}
-
 @article{penn2024,
   title={Phylo2Vec: a vector representation for binary trees},
   author={Penn, Matthew J and Scheidwasser, Neil and Khurana, Mark P and Duch{\^e}ne, David A and Donnelly, Christl A and Bhatt, Samir},
