+++ joss/paper.md
@@ -56,28 +56,28 @@ bibliography: paper.bib
 
 # Summary
 
-Phylogenetics is a fundamental component of many analysis frameworks in biology [@yang2014] as well as linguistics [@atkinson2005] to study the evolutionary relationships of different entities. Recently, the advent of large-scale genomics and the SARS-CoV-2 pandemic has underscored the necessity for phylogenetic software to handle large datasets of genomes or phylogenetic trees [@kapli2020; @attwood2022; @khurana2024; @kraemer2025]. While significant efforts have focused on scaling optimisation algorithms [@turakhia2021; @sanderson2021; @demaio2023], visualization [@sanderson2022], and lineage identification [@mcbroome2024], an emerging body of research has been dedicated to efficient representations of data for genomes [@deorowicz2023] and phylogenetic trees such as phylo2vec [@penn2024; @chauve2025; @richman2025]. Compared to traditional tree representations such as the Newick format [@felsenstein2004], which represents trees using strings of nested parentheses, modern representations of phylogenetic trees utilize integer vectors to define the tree topology traversal. This approach offers several advantages, including easier manipulability, increased memory efficiency, and applicability to downstream tasks such as machine learning.
+Phylogenetics is a fundamental component of evolutionary analysis frameworks in biology [@yang2014] and linguistics [@atkinson2005]. Recently, the advent of large-scale genomics and the SARS-CoV-2 pandemic has highlighted the necessity for phylogenetic software to handle large datasets [@kapli2020; @attwood2022; @khurana2024; @kraemer2025]. While significant efforts have focused on scaling optimisation algorithms [@turakhia2021; @sanderson2021; @demaio2023], visualization [@sanderson2022], and lineage identification [@mcbroome2024], an emerging body of research has been dedicated to efficient representations of data for genomes [@deorowicz2023] and phylogenetic trees such as phylo2vec [@penn2024; @chauve2025; @richman2025]. Compared to the traditional Newick format which represents trees using strings of nested parentheses [@felsenstein2004], modern tree representations utilize integer vectors to define the tree topology traversal. This approach offers several advantages, including easier manipulation, increased memory efficiency, and applicability to machine learning.
 
-Here, we present the latest release of \texttt{phylo2vec} (or Phylo2Vec), a high-performance software package for encoding, manipulating, and analysing binary phylogenetic trees. At its core, the package is based on the phylo2vec [@penn2024] representation of binary trees, which defines a bijection from any tree topology with $n$ leaves into an integer vector of size $n-1$. Compared to the traditional Newick format, phylo2vec is designed to enable fast sampling and comparison of binary trees. This release features a core implementation in Rust, providing significant performance improvements and memory efficiency (\autoref{fig:benchmarks}), while remaining available in Python (superseding the release described in the original paper [@penn2024]) and R via dedicated wrappers, making it accessible to a broad audience in the bioinformatics community.
+Here, we present the latest release of \texttt{phylo2vec} (or Phylo2Vec), a high-performance software package for encoding, manipulating, and analysing binary phylogenetic trees. At its core, the package is based on the phylo2vec [@penn2024] representation of binary trees, and is designed to enable fast sampling and tree comparison. This release features a core implementation in Rust for improved performance and memory efficiency (\autoref{fig:benchmarks}), with wrappers in R and Python (superseding the original release [@penn2024]), making it accessible to a broad audience in the bioinformatics community.
 
[…]
