### 2025-09-16 09bd89e
feat(r): add remove_branch_lengths and remove_parent_labels to R interface
