+++ r-phylo2vec/man/remove_parent_labels.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{remove_parent_labels}
+\alias{remove_parent_labels}
+\title{Remove parent labels from the Newick string}
+\usage{
+remove_parent_labels(newick)
+}
+\description{
+Remove parent labels from the Newick string
+}
