+++ r-phylo2vec/NAMESPACE
@@ -25,7 +25,9 @@ export(pre_precision_from_matrix)
 export(pre_precision_from_vector)
 export(precision)
 export(queue_shuffle)
+export(remove_branch_lengths)
 export(remove_leaf)
+export(remove_parent_labels)
 export(sample_matrix)
 export(sample_tree)
 export(sample_vector)
