### 2025-09-17 #8407 @richelbilderbeek
Hi @abhishektiwari (and team), this seems quite useful what you've done here! I **accept** to review this, with a focus on the R wrapper parts :+1:
