+++ py-phylo2vec/phylo2vec/utils/vector.py
@@ -7,7 +7,7 @@ from typing import List, Tuple
 
 import numpy as np
 
-from ete3 import Tree
+from ete4 import Tree
 
 from phylo2vec import _phylo2vec_core as core
 from phylo2vec.base.newick import from_newick, to_newick
@@ -273,11 +273,11 @@ def reroot(v, node) -> np.ndarray:
     numpy.ndarray
         rerooted vector
     """
-    ete3_tree = Tree(to_newick(v), format=8)
+    ete_tree = Tree(to_newick(v), parser=8)
 
-    ete3_tree.set_outgroup(f"{node}")
+    ete_tree.set_outgroup(f"{node}")
 
-    newick = ete3_tree.write(format=9)
+    newick = ete_tree.write(parser=9)
 
     v_new = from_newick(newick)
 
