+++ py-phylo2vec/pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
   "numpy>=1.22,<2.1",
   "biopython>=1.84.0",
   "joblib>=1.2.0",
-  "ete3==3.1.3",
+  "ete4==4.4.0",
   "six>=1.16.0",
 ]
 
