### 2025-09-19 54c521d
fix: remove duplicate ete4 dependency
