+++ py-phylo2vec/phylo2vec/utils/vector.py
@@ -8,7 +8,6 @@ from typing import List, Tuple
 import numpy as np
 
 from ete4 import Tree
-from ete4 import Tree
 
 from phylo2vec import _phylo2vec_core as core
 from phylo2vec.base.newick import from_newick, to_newick
