### 2025-09-19 acbe0b4
fix: add libcrypt to python deps
