### 2025-09-19 4cd3612
Update ete3 to ete4 + add support for Python 3.13 (#108)
* refactor: update ete3 --> ete4

* fix: remove duplicate ete4 dependency

* refactor: rename ete3 to ete4 in markdown cells of the demo

* fix: add libcrypt to python deps

* feat: add py313 to ci
