+++ py-phylo2vec/phylo2vec/opt/utils.py
@@ -1,3 +1,7 @@
+import logging
+
+logging.getLogger("rpy2").setLevel(logging.ERROR)
+
 import rpy2
 import rpy2.robjects as ro
 
@@ -5,8 +9,9 @@ from rpy2.robjects import pandas2ri
 from rpy2.robjects.conversion import localconverter
 from rpy2.robjects.packages import importr
 
-# Disable rpy2 warning
+# Disable rpy2 warnings
 rpy2.rinterface_lib.callbacks.consolewrite_warnerror = lambda *args: None
+rpy2.rinterface_lib.init_ = ["--quiet", "--no-save"]
 
 
 def dnadist(fasta_path, model, gamma=False):
