+++ .github/workflows/ci-python.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-environment: ['py310', 'py311', 'py312']
+        python-environment: ['py310', 'py311', 'py312', 'py313']
     steps:
       - uses: actions/checkout@v5
       - uses: prefix-dev/setup-pixi@v0.9.0
