+++ py-phylo2vec/pyproject.toml
@@ -10,7 +10,7 @@ authors = [
   { name = "Neil Scheidwasser", email = "neil.clow@sund.ku.dk" }
 ]
 readme = "README.md"
-requires-python = ">=3.10,<3.13"
+requires-python = ">=3.10,<=3.13"
 license = {file = "LICENSE"}
 classifiers = [
     "Development Status :: 3 - Alpha",
@@ -20,6 +20,7 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
     "Operating System :: Unix",
     "Operating System :: Microsoft :: Windows",
@@ -28,7 +29,7 @@ dependencies = [
   "numpy>=1.22,<2.1",
   "biopython>=1.84.0",
   "joblib>=1.2.0",
-  "ete3==3.1.3",
+  "ete4==4.4.0",
   "six>=1.16.0",
 ]
 
