### 2025-09-19 3c6edfa
fix: add missing update to newick_patterns
