+++ phylo2vec/src/newick/newick_patterns.rs
@@ -17,30 +17,48 @@
 #[derive(Debug)]
 pub struct NewickPatterns {
     pub left_node: regex::Regex,
+    pub left_node_generic: regex::Regex,
     pub right_node: regex::Regex,
+    pub right_node_generic: regex::Regex,
     pub pairs: regex::Regex,
     pub branch_lengths: regex::Regex,
     pub parents: regex::Regex,
 }
 
 impl NewickPatterns {
+    /// Create a new instance of `NewickPatterns`.
+    ///
+    /// Patterns:
+    /// * `left_node`: `\(\b(\d+)\b`
+    /// * `right_node`: `,\b(\d+)\b`
+    /// * `pairs`: `({})|({})`
+    /// * `branch_lengths`: `:\d+(\.\d+)?`
+    /// * `parents`: `\)(\d+)`
+    /// # Panics
+    /// This function will panic if the regular expressions are invalid.
     pub fn new() -> Self {
-        let _left_node = r"\(\b(\d+)\b";
-        let _right_node = r",\b(\d+)\b";
-        let _branch_lengths = r":\d+(\.\d+)?";
-        let _parents = r"\)(\d+)";
-        let _pairs = format!(r"({})|({})", _left_node, _right_node);
-        NewickPatterns {
+        let lnode_pattern = r"\(\b(\d+)\b";
+        let rnode_pattern = r",\b(\d+)\b";
+        let pnode_pattern = r"\)(\d+)";
+        let lnode_generic_pattern = r"\(\b(\w+)\b";
+        let rnode_generic_pattern = r",\b(\w+)\b";
+        let bl_pattern = r":\d+(\.\d+)?";
+        let pair_pattern = format!(r"({lnode_pattern})|({rnode_pattern})");
+        Self {
             // Pattern of an integer label on the left of a pair
-            left_node: regex::Regex::new(_left_node).unwrap(),
+            left_node: regex::Regex::new(lnode_pattern).unwrap(),
             // Pattern of an integer label on the right of a pair
-            right_node: regex::Regex::new(_right_node).unwrap(),
+            right_node: regex::Regex::new(rnode_pattern).unwrap(),
+            // Pattern of a generic label on the left of a pair
+            left_node_generic: regex::Regex::new(lnode_generic_pattern).unwrap(),
+            // Pattern of a generic label on the right of a pair
+            right_node_generic: regex::Regex::new(rnode_generic_pattern).unwrap(),
             // Pattern of a pair of integer labels
-            pairs: regex::Regex::new(&_pairs).unwrap(),
+            pairs: regex::Regex::new(&pair_pattern).unwrap(),
             // Pattern of a branch length annotation
-            branch_lengths: regex::Regex::new(_branch_lengths).unwrap(),
+            branch_lengths: regex::Regex::new(bl_pattern).unwrap(),
             // Pattern of a parent label
-            parents: regex::Regex::new(_parents).unwrap(),
+            parents: regex::Regex::new(pnode_pattern).unwrap(),
         }
     }
 }
