### 2025-09-21 #8407 @abhishektiwari
Hi @theosanderson, would you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "phylo2vec: a library for vector-based phylogenetic tree manipulation".

Based on your experience with Phylogenetics, we think you might be able to provide a great review of this submission.

Thanks
Abhi
