### 2025-09-21 #8407 @abhishektiwari
Hi @amcandio thanks. Adding you specifically to cover the Rust core, associated performance claims, and Python interfaces.
