### 2025-09-21 #9040 @editorialbot
[REVIEW]: phylo2vec: a library for vector-based phylogenetic tree manipulation
**Submitting author:** @Neclow (<a href="http://orcid.org/0000-0001-9922-0289">Neil Alexandre Scheidwasser</a>)
**Repository:** https://github.com/sbhattlab/phylo2vec
**Branch with paper.md** (empty if default branch): main
**Version:** v1.5.0
