### 2025-09-22 sbhattlab/phylo2vec#113 @tkchafin
[JOSS Review] Rename package files in release assets
The README states to use the appropriate package attached to the release for your OS, but it isn't clear which one to use since they are not all named with the OS/architecture

`cd-r.yaml` makes some packages named like `dist/r-package*.zip `, but then these are uploaded to the release assets as:
- phylo2vec_1.4.0.tgz (macOS)
- phylo2vec_1.4.0.zip (windows)
- phylo2vec_1.4.0_R_x86_64-pc-linux-gnu.tar.gz (linux)

It would be easier for people if you named them with e.g., `_macos-arm64.tar.gz` etc

https://github.com/openjournals/joss-reviews/issues/9040
