### 2025-09-22 361d8fc
chore: update pre-commit hooks
updates:
- [github.com/python-jsonschema/check-jsonschema: 0.33.3 → 0.34.0](https://github.com/python-jsonschema/check-jsonschema/compare/0.33.3...0.34.0)
