### 2025-09-22 8db53c8
fix(docs): fix broken readme link + add more contrib guidelines
