+++ CONTRIBUTING.md
@@ -1,12 +1,39 @@
-## Contributing to Phylo2Vec
+# Contributing to Phylo2Vec
 
 Thank you for your interest in contributing to Phylo2Vec! We welcome
 contributions from the community to help improve our project.
 
+## Getting started
+
 To get started, please refer to our
-[online documentation](https://phylo2vec.readthedocs.io/en/latest/development.html)
-and [README](https://github.com/phylo2vec?tab=readme-ov-file#phylo2vec) for
-detailed guidelines on how to contribute.
+[Development guide](https://phylo2vec.readthedocs.io/en/latest/development.html) to familiarize yourself with our project’s structure, testing framework, and benchmarking practices.
+
+## Reporting Bugs and Requesting Features
+
+We use GitHub Issues to track both bugs and feature requests.
+
+If you encounter a bug or would like to suggest a new feature, please open a new issue on the [Issues page](https://github.com/sbhattlab/phylo2vec/issues).
+
+To help us handle issue swiftly, please include the following details:
+
+* **Description**: A concise explanation of the problem or request, along with the expected outcome.
+* **For bugs**: A minimal reproducible example and the complete error message or stack trace.
+* **Environment details**: Your operating system and, if applicable, a list of relevant dependencies. Examples:
+  * Python: NumPy, Biopython, pytest, ete
+  * Rust: rand, ndarray, regex
+  * R: ape, testthat
+
+## Contributing Code
+
+If you would like to contribute code directly, please follow these steps:
+
+1. **Fork the repository** and create a branch from `main`:
+
+```bash
+git clone https://github.com/<your-username>/phylo2vec.git
+cd phylo2vec
+git checkout -b my-feature-branch
+```
 
 We appreciate your contributions and look forward to your involvement with this
 project!
