### 2025-09-23 c8099a1
docs: finalize CONTRIBUTING pr steps
