### 2025-09-23 558093d
style: pre-commit fixes
