### 2025-09-23 d1c52b1
docs: add links to CoC and README in Contributing; add LICENSE link to README
