### 2025-09-23 9351c2c
style: pre-commit fixes
