### 2025-09-23 90dd145
docs: Update contributing guidelines and API reference (#116)
* fix(docs): fix broken readme link + add more contrib guidelines

* docs: finalize CONTRIBUTING pr steps

* style: pre-commit fixes

* docs: add links to CoC and README in Contributing; add LICENSE link to README

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
