### 2025-09-23 f74f5b0
chore: update R CD to add OSs in release names
