### 2025-09-23 afcdf93
chore: update R CD to add OSs in release names (#117)
* chore: update R CD to add OSs in release names

* chore: update mv command for win R package
