### 2025-09-23 73cd46a
feat: add more versatile newick processing
