+++ phylo2vec/src/newick/mod.rs
@@ -261,13 +261,22 @@ pub fn parse_with_bls(newick: &str) -> Result<(Ancestry, Vec<[f64; 2]>), NewickE
 /// ```
 /// use phylo2vec::newick::remove_parent_labels;
 ///
+/// // Integer Newick
 /// let newick = "(((0,(3,5)6)8,2)9,(1,4)7)10;";
 /// let result = remove_parent_labels(newick);
 /// assert_eq!(result, "(((0,(3,5)),2),(1,4));");
+///
+/// // String Newick
+/// let newick_str = "(((A,(C,E)p0)p1,B)p2,(D,F)p3)p4;";
+/// let result_str = remove_parent_labels(newick_str);
+/// assert_eq!(result_str, "(((A,(C,E)),B),(D,F));");
 /// ```
 pub fn remove_parent_labels(newick: &str) -> String {
     let newick_patterns = NewickPatterns::new();
-    newick_patterns.parents.replace_all(newick, ")").to_string()
+    newick_patterns
+        .parent_generic
+        .replace_all(newick, ")")
+        .to_string()
 }
 
 /// Remove branch length annotations from the Newick string
@@ -306,7 +315,7 @@ pub fn remove_branch_lengths(newick: &str) -> String {
 /// ```
 pub fn has_parents(newick: &str) -> bool {
     let newick_patterns = NewickPatterns::new();
-    newick_patterns.parents.is_match(newick)
+    newick_patterns.parent.is_match(newick)
 }
 
 /// Check if the Newick string has branch lengths
@@ -328,25 +337,34 @@ pub fn has_branch_lengths(newick: &str) -> bool {
     newick_patterns.branch_lengths.is_match(newick)
 }
 
-/// Check if the Newick string has integer labels
+/// Check if a Newick string has only integer leaves
 ///
 /// # Examples
 ///
 /// ```
-/// use phylo2vec::newick::has_integer_labels;
+/// use phylo2vec::newick::has_integer_leaves;
+///
+/// // Only string leaves
 /// let nw_str = "(((E:0.1,(A:0.1,C:0.1):0.1):0.1,F:0.1):0.1,(D:0.1,B:0.1):0.1);";
-/// let result_str = has_integer_labels(nw_str);
+/// let result_str = has_integer_leaves(nw_str);
 /// assert!(!result_str);
 ///
+/// // Mixed string and integer leaves
 /// let nw_mixed = "(((0:0.1,(A:0.1,C:0.1)6:0.1)8:0.1,2:0.1)9:0.1,(1:0.1,4:0.1)7:0.1)10;";
-/// let result_mixed = has_integer_labels(nw_mixed);
+/// let result_mixed = has_integer_leaves(nw_mixed);
 /// assert!(!result_mixed);
 ///
+/// // Only integer leaves
 /// let nw_int = "(((0:0.1,(3:0.1,5:0.1)6:0.1)8:0.1,2:0.1)9:0.1,(1:0.1,4:0.1)7:0.1)10;";
-/// let result_int = has_integer_labels(nw_int);
+/// let result_int = has_integer_leaves(nw_int);
 /// assert!(result_int);
+///
+/// // Integer leaves and string parent nodes
+/// let nw_int_parents = "(((0:0.1,(3:0.1,5:0.1)p0:0.1)p1:0.1,2:0.1)p2:0.1,(1:0.1,4:0.1)p3:0.1)p4;";
+/// let result_int_parents = has_integer_leaves(nw_int_parents);
+/// assert!(result_int_parents);
 /// ```
-pub fn has_integer_labels(newick: &str) -> bool {
+pub fn has_integer_leaves(newick: &str) -> bool {
     let newick_patterns = NewickPatterns::new();
[…]
