+++ r-phylo2vec/man/process_newick.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{process_newick}
+\alias{process_newick}
+\title{Process a Newick string}
+\usage{
+process_newick(newick)
+}
+\description{
+Process a Newick string
+}

[…]
