### 2025-09-23 0d96f5d
feat(python): add load_newick to cli
