### 2025-09-23 9367b7d
fix(python): use print vs. pprint to pass tests
