### 2025-09-27 sbhattlab/phylo2vec#120 @amcandio
[JOSS Review] guideline for reporting issues and seek support
Hi! It would be great to clarify what are the guidelines to report issues and seek support. Is github issues the recommended way?

https://github.com/openjournals/joss-reviews/issues/9040
