### 2025-09-27 sbhattlab/phylo2vec#120 @amcandio
Ah nevermind, I see they are in CONTRIBUTING. I think it would be better to have the issue reporting in the README. This is optional so closing the issue.
