### 2025-09-27 sbhattlab/phylo2vec#121 @amcandio
[JOSS Review] Manuscript comments
Hi! Here are some comments on the manuscript:

- Typo "ran on an on an AMD" -> "ran on an". Or "Ran on a computer with processor AMD ..." would be clearer
- I think the feature section is great but it can be too detailed for JOSS scope. I'd recommend simplifying that section with just a mention to the tree annotation feature without having to explain it in detail. You can refer to the original paper or to the library documentation for that.

https://github.com/openjournals/joss-reviews/issues/9040
