### 2025-09-29 278bf69
docs(paper): clarify paragraphs in paper + add compilation guidelines
