+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @article{penn2024,
   title={Phylo2Vec: a vector representation for binary trees},
   author={Penn, Matthew J and Scheidwasser, Neil and Khurana, Mark P and Duch{\^e}ne, David A and Donnelly, Christl A and Bhatt, Samir},
-  journal={Systematic Biology},
+  journal={Syst. Biol.},
   pages={syae030},
   year={2024},
   doi={10.1093/sysbio/syae030},
@@ -34,7 +34,7 @@
 @article{ayres2012,
   title={BEAGLE: an application programming interface and high-performance computing library for statistical phylogenetics},
   author={Ayres, Daniel L and Darling, Aaron and Zwickl, Derrick J and Beerli, Peter and Holder, Mark T and Lewis, Paul O and Huelsenbeck, John P and Ronquist, Fredrik and Swofford, David L and Cummings, Michael P and others},
-  journal={Systematic biology},
+  journal={Syst. Biol.},
   volume={61},
   number={1},
   pages={170--173},
@@ -49,13 +49,14 @@
   journal={Proc. USSR Acad. Sci.},
   volume={146},
   pages={263--266},
-  year={In Russian. English translation in Soviet Mathematics Doklady, 3:1259--1263, 1962}
+  year={1962},
+  note={In Russian. English translation in Soviet Mathematics Doklady, 3:1259--1263}
 }
 
 @article{penn2023,
   title={Leaping through tree space: continuous phylogenetic inference for rooted and unrooted trees},
   author={Penn, Matthew J and Scheidwasser, Neil and Penn, Joseph and Donnelly, Christl A and Duch{\^e}ne, David A and Bhatt, Samir},
-  journal={Genome Biology and Evolution},
+  journal={Genome Biol. Evol.},
   volume={15},
   number={12},
   pages={evad213},
@@ -65,7 +66,7 @@
 }
 
 @article{duchemin2018,
-    author = {Duchemin, Wandrille and Gence, Guillaume and Arigon Chifolleau, Anne-Muriel and Arvestad, Lars and Bansal, Mukul S and Berry, Vincent and Boussau, Bastien and Chevenet, François and Comte, Nicolas and Davín, Adrián A and Dessimoz, Christophe and Dylus, David and Hasic, Damir and Mallo, Diego and Planel, Rémi and Posada, David and Scornavacca, Celine and Szöllősi, Gergely and Zhang, Louxin and Tannier, Éric and Daubin, Vincent},
+    author = {Duchemin, Wandrille and Gence, Guillaume and Arigon Chifolleau, Anne-Muriel and Arvestad, Lars and Bansal, Mukul S and Berry, Vincent and Boussau, Bastien and Chevenet, Fran{\c{c}}ois and Comte, Nicolas and Dav{\'\i}n, Adri{\'a}n A and others},
     title = {RecPhyloXML: a format for reconciled gene trees},
     journal = {Bioinformatics},
     volume = {34},
@@ -80,7 +81,7 @@
 @article{nakhleh2013,
   title={Computational approaches to species phylogeny inference and gene tree reconciliation},
   author={Nakhleh, Luay},
-  journal={Trends in ecology \& evolution},
+  journal={Trends Ecol. Evol.},
   volume={28},
   number={12},
   pages={719--728},
@@ -104,7 +105,7 @@
 @article{huerta2016,
   title={ETE 3: reconstruction, analysis, and visualization of phylogenomic data},
   author={Huerta-Cepas, Jaime and Serra, Fran{\c{c}}ois and Bork, Peer},
-  journal={Molecular biology and evolution},
+  journal={Mol. Biol. Evol.},
   volume={33},
   number={6},
[…]
