+++ joss/paper.md
@@ -56,12 +56,11 @@ bibliography: paper.bib
 
 # Summary
 
-Phylogenetics is a fundamental component of evolutionary analysis frameworks in biology [@yang2014] and linguistics [@atkinson2005]. Recently, the advent of large-scale genomics and the SARS-CoV-2 pandemic has highlighted the necessity for phylogenetic software to handle large datasets [@kapli2020; @attwood2022; @khurana2024; @kraemer2025]. While significant efforts have focused on scaling optimisation algorithms [@turakhia2021; @sanderson2021; @demaio2023], visualization [@sanderson2022], and lineage identification [@mcbroome2024], an emerging body of research has been dedicated to efficient representations of data for genomes [@deorowicz2023] and phylogenetic trees such as phylo2vec [@penn2024; @chauve2025; @richman2025]. Compared to the traditional Newick format which represents trees using strings of nested parentheses [@felsenstein2004], modern tree representations utilize integer vectors to define the tree topology traversal. This approach offers several advantages, including easier manipulation, increased memory efficiency, and applicability to machine learning.
+Phylogenetics is a fundamental component of evolutionary analysis frameworks in biology [@yang2014] and linguistics [@atkinson2005]. Recently, the advent of large-scale genomics and the SARS-CoV-2 pandemic has highlighted the necessity for phylogenetic software to handle large datasets [@kapli2020; @attwood2022; @khurana2024; @kraemer2025]. While significant efforts have focused on scaling optimisation algorithms [@turakhia2021; @sanderson2021; @demaio2023], visualization [@sanderson2022], and lineage identification [@mcbroome2024], an emerging body of research has been dedicated to efficient representations of data for genomes [@deorowicz2023] and phylogenetic trees [@penn2024; @chauve2025; @richman2025]. Compared to the traditional Newick format which represents trees using strings of nested parentheses [@felsenstein2004], modern tree representations utilize integer vectors to define the tree topology traversal. This approach offers several advantages, including easier manipulation, increased memory efficiency, and applicability to machine learning.
 
 Here, we present the latest release of \texttt{phylo2vec} (or Phylo2Vec), a high-performance software package for encoding, manipulating, and analysing binary phylogenetic trees. At its core, the package is based on the phylo2vec [@penn2024] representation of binary trees, and is designed to enable fast sampling and tree comparison. This release features a core implementation in Rust for improved performance and memory efficiency (\autoref{fig:benchmarks}), with wrappers in R and Python (superseding the original release [@penn2024]), making it accessible to a broad audience in the bioinformatics community.
 
-![Benchmark times for converting a phylo2vec vector to a Newick string (left) and vice versa (right). Execution time was measured over at least 20 runs per size, comparing Python functions in the latest release (via Rust bindings with [PyO3](https://github.com/PyO3/pyo3)) against the previous release [@penn2024] based on \texttt{Numba} [@lam2015]. Benchmarks ran on an
-on an AMD Ryzen Threadripper PRO 5995WX (64 cores, 7 GHz, 256 GB RAM). \label{fig:benchmarks}](fig1.pdf)
+![Benchmark times for converting a phylo2vec vector to a Newick string (left) and vice versa (right). Execution time was measured over at least 20 runs per size, comparing Python functions in the latest release (via Rust bindings with [PyO3](https://github.com/PyO3/pyo3)) against the previous release [@penn2024] based on \texttt{Numba} [@lam2015]. All benchmarks were performed on a workstation equipped with an AMD Ryzen Threadripper PRO 5995WX (64 cores, 2.7 GHz) and 256 GB of RAM. \label{fig:benchmarks}](fig1.pdf)
 
[…]
