### 2025-09-29 sbhattlab/phylo2vec#121 @Neclow
Thanks for the comment! I wasn't entirely sure what you meant by "mention to the tree annotation feature without having to explain it in detail", but I have tried to simplify the feature section, especially related to Fig. 2. See the upcoming PR for more details
