### 2025-09-29 fab8c46
fix(ci): add rustfmt install for rust 1.90 (#124)
