+++ .github/workflows/ci-rust.yaml
@@ -26,6 +26,8 @@ jobs:
       - uses: actions/checkout@v5
       - name: Setup Rust
         run: rustup update ${{ matrix.toolchain }} && rustup default ${{ matrix.toolchain }}
+      - name: Install rustfmt
+        run: rustup component add rustfmt
       - name: Run linter
         run: cargo fmt --check
       - name: Compile Code
