### 2025-09-29 15ad906
Update joss/src/wordcount.rb
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
