### 2025-09-29 df35c2a
docs(paper): clarify paragraphs in paper + add compilation guidelines (#122)
* docs(paper): clarify paragraphs in paper + add compilation guidelines

* Update joss/src/wordcount.rb

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
