### 2025-09-29 2ec4150
docs(readme): add small subsection in README for reporting issues
