### 2025-09-29 a22c3ee
style: pre-commit fixes
