### 2025-09-29 f8ce7a2
docs(readme): fix url for pulls
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
