### 2025-09-29 b0d038b
style: pre-commit fixes
