### 2025-09-29 33c760f
fix(docs): fix copilots quirks
