### 2025-09-29 c8b0924
style: pre-commit fixes
