### 2025-09-29 c65ea23
docs(readme): add small subsection in README for reporting issues (#125)
* docs(readme): add small subsection in README for reporting issues

* style: pre-commit fixes

* docs(readme): fix url for pulls

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* style: pre-commit fixes

* fix(docs): fix copilots quirks

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
