### 2025-09-29 sbhattlab/phylo2vec#121 @amcandio
Thanks! What I meant about Figure 2 is that it is out of the scope of the paper, according to JOSS guidelines https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain.

> Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

So, a reference to the paper or documentation would suffice
