### 2025-09-29 sbhattlab/phylo2vec#121 @Neclow
Sorry for the misunderstanding! I will remove figure 2. As it's a new format developed after the publication of the original phylo2vec paper, I will reference the documentation.
