### 2025-09-29 c14c3c5
docs(paper): move fig2 from joss to demo (#127)
