+++ joss/paper.md
@@ -70,9 +70,7 @@ The purpose of the \texttt{phylo2vec} library is threefold. First, it provides r
 
 The presented release of \texttt{phylo2vec} addresses optimisations limitations of [@penn2024] with $\mathcal{O}(n \log n)$ implementations for vector-to-Newick and Newick-to-vector conversions, leveraging Adelson-Velsky and Landis (AVL) trees [@adelson1962] and Fenwick trees [@fenwick1994], respectively.
 
-New features include an extension of the vector representation to support branch length annotations (\autoref{fig:phylo2mat}), leaf-level operations (pruning, placement, MRCA identification), fast cophenetic distance matrix calculation, and various optimisation schemes based on phylo2vec tree representations, notably hill-climbing [@penn2024] and GradME [@penn2023]. We also propose a likelihood function for Bayesian MCMC inference that leverages tree representation similarities with \texttt{BEAGLE} [@suchard2009; @ayres2012]. Finally, user-friendliness is enhanced with step-by-step demos of phylo2vec’s representations and core functions.
-
-![Recovering a tree from a phylo2vec vector: example for $\boldsymbol{v} = [0, 2, 0, 4]$. (a) Leaf placement algorithm from [@penn2024]. (b) Augmenting $\boldsymbol{v}$ into a matrix $\boldsymbol{m}$ with branch lengths. We use an intermediary ancestry matrix whereby each row encodes a cherry (two children + parent) with two extra columns for branch lengths. The node with the smallest descendant has the branch length in the second column, and the other the branch length in the third column.\label{fig:phylo2mat}](fig2.pdf)
+New features include an extension of the vector representation to support branch length annotations, leaf-level operations (pruning, placement, MRCA identification), fast cophenetic distance matrix calculation, and various optimisation schemes based on phylo2vec tree representations, notably hill-climbing [@penn2024] and GradME [@penn2023]. We also propose a likelihood function for Bayesian MCMC inference that leverages tree representation similarities with \texttt{BEAGLE} [@suchard2009; @ayres2012]. Finally, user-friendliness is enhanced with step-by-step demos of phylo2vec’s representations and core functions.
 
 # Maintenance
 
