### 2025-10-02 f2c9651
docs(paper): update author list (#128)
