### 2025-10-03 2b4b607
docs(data): add download links for each dataset + data section in readme
