+++ py-phylo2vec/phylo2vec/datasets/descr/zika.md
@@ -1,8 +1,8 @@
 # Zika dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description                                                                                                       |
-| -------------- | -------------- | --------------- | ---- | ----------------------------------------------------------------------------------------------------------------- |
-| Zika virus     | 86             | 10,807          | RNA  | Zika whole genomes isolated between 2013 and 2016 (Retrieved from https://github.com/neherlab/treetime_examples). |
+| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
+| -------------- | -------------- | --------------- | ---- |------------ | ---- |
+| Zika virus     | 86             | 10,807          | RNA  | Zika whole genomes isolated between 2013 and 2016. Retrieved from [treetime](https://github.com/neherlab/treetime_examples). | [FASTA Link](https://github.com/neherlab/treetime_examples/blob/79eae7f8025a8ef3165c856c7359d92e738eb893/data/zika/zika.fasta) |
 
 References:
 
