+++ py-phylo2vec/phylo2vec/datasets/descr/m501.md
@@ -1,12 +1,12 @@
-## M501 dataset
+# M501 dataset
 
-| Taxonomic Rank   | Number of Taxa | Number of Bases | Type | Description                                                                                                                 |
-| ---------------- | -------------- | --------------- | ---- | --------------------------------------------------------------------------------------------------------------------------- |
-| Acanthocephalans | 29             | 2,520           | RNA  | 18S rRNA sequences of Acanthocephala (thorny-headed worms) and Rotifera species (Retrieved from https://www.treebase.org/). |
+| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
+| -------------- | -------------- | --------------- | ---- |------------ | ---- |
+| Acanthocephalans | 29             | 2,520           | RNA  | 18S rRNA sequences of Acanthocephala (thorny-headed worms) and Rotifera species. Originally retrieved from [treebase](https://www.treebase.org). | [Nexus link](https://datadryad.org/downloads/file_stream/84476) |
 
 References:
 
-```
+```latex
 @ARTICLE{Garey1996-jc,
   title    = "Molecular evidence for Acanthocephala as a subtaxon of Rotifera",
   author   = "Garey, J R and Near, T J and Nonnemacher, M R and Nadler, S A",
@@ -15,6 +15,5 @@ References:
   number   =  3,
   pages    = "287--292",
   year     =  1996,
-
 }
 ```
