### 2025-10-03 ba8fd76
style: pre-commit fixes
