+++ py-phylo2vec/phylo2vec/datasets/descr/m501.md
@@ -1,7 +1,7 @@
 # M501 dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
-| -------------- | -------------- | --------------- | ---- |------------ | ---- |
+| Taxonomic Rank   | Number of Taxa | Number of Bases | Type | Description                                                                                                                                      | Link                                                            |
+| ---------------- | -------------- | --------------- | ---- | ------------------------------------------------------------------------------------------------------------------------------------------------ | --------------------------------------------------------------- |
 | Acanthocephalans | 29             | 2,520           | RNA  | 18S rRNA sequences of Acanthocephala (thorny-headed worms) and Rotifera species. Originally retrieved from [treebase](https://www.treebase.org). | [Nexus link](https://datadryad.org/downloads/file_stream/84476) |
 
 References:
