### 2025-10-03 2244776
docs(data): add download links for each dataset + data section in readme (#130)
* docs(data): add download links for each dataset + data section in readme

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
