+++ py-phylo2vec/phylo2vec/datasets/descr/yeast.md
@@ -1,8 +1,8 @@
 # Yeast dataset
 
-| Taxonomic Rank                | Number of Taxa | Number of Bases | Type | Description                                                                                                                 |
-| ----------------------------- | -------------- | --------------- | ---- | --------------------------------------------------------------------------------------------------------------------------- |
-| Fungi; mainly _Saccharomyces_ | 8              | 127,018         | DNA  | Alignment of 106 orthologous genes of 8 different species of yeast (Retrieved from https://github.com/KlausVigo/phangorn/). |
+| Taxonomic Rank                | Number of Taxa | Number of Bases | Type | Description                                                                                                                            | Link                                                                           |
+| ----------------------------- | -------------- | --------------- | ---- | -------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------ |
+| Fungi; mainly _Saccharomyces_ | 8              | 127,018         | DNA  | Alignment of 106 orthologous genes of 8 different species of yeast. Retrieved from [phangorn](https://github.com/KlausVigo/phangorn/). | [RData link](https://github.com/KlausVigo/phangorn/blob/main/data/yeast.RData) |
 
 References:
 
