### 2025-10-07 2eabcd0
data(py): add alignment loading for penn2023 datasets + tests + docs
