+++ py-phylo2vec/phylo2vec/datasets/_base.py
@@ -1,11 +1,41 @@
 """Base IO code for all datasets."""
 
+import json
+import os
+import re
+
+
 from importlib import resources
+from pathlib import Path
+from urllib.parse import (
+    urlparse,
+    uses_netloc,
+    uses_params,
+    uses_relative,
+)
+from urllib.request import urlretrieve
 
 from Bio import SeqIO
 
 DATA_MODULE = "phylo2vec.datasets.data"
 DESCR_MODULE = "phylo2vec.datasets.descr"
+DATASETS = {
+    "fluA": {
+        "format": "fasta",
+    },
+    "h3n2": {
+        "format": "fasta",
+    },
+    "m501": {
+        "format": "fasta",
+    },
+    "yeast": {
+        "format": "RData",
+    },
+    "zika": {
+        "format": "fasta",
+    },
+}
 
 
 def read_fasta(data_path):
@@ -44,7 +74,120 @@ def load_fasta(data_file_name, data_module=DATA_MODULE):
     return read_fasta(_open_text(data_module, data_file_name))
 
 
-def load_descr(descr_file_name, descr_module=DESCR_MODULE):
+FORMATS = {
+    "fasta": {
+        "ext": "fa",
+        "loader": load_fasta,
+    },
+}
+
+
+def list_datasets():
+    """List all available datasets."""
+    # Check that format is supported
+    return [dataset for dataset in DATASETS if DATASETS[dataset]["format"] in FORMATS]
+
+
+def is_url(maybe_url: str) -> bool:
+    # pylint: disable=line-too-long
+    """Check if a string is a URL.
+
+    Adapted from pandas.
+    Source: https://github.com/pandas-dev/pandas/blob/c8213d16d98079da2a7b7464b95fe22f7d72d427/pandas/io/common.py
+    # Copyright (c) 2008-2011, AQR Capital Management, LLC, Lambda Foundry, Inc. and PyData Development Team
+    # Copyright (c) 2011-2025, Open source contributors.
+    # BSD 3-Clause License
+
+
+    Parameters
+    ----------
+    maybe_url : str
+        String to check
+    Returns
+    -------
+    bool
+        True if the string is a URL, False otherwise
+    """
+    # pylint: enable=line-too-long
+    _VALID_URLS = set(uses_relative + uses_netloc + uses_params)
+    _VALID_URLS.discard("")
+
+    if not isinstance(maybe_url, str):
+        return False
+    return urlparse(maybe_url).scheme in _VALID_URLS
+
+
+def load_alignment(
+    dataset,
+    data_format=None,
+    data_module=DATA_MODULE,
+    descr_module=DESCR_MODULE,
+    url_key="Link",
+    overwrite=False,
+):
+    if dataset in DATASETS:
+        # Load description
+        print(f"Loading {dataset}...")
+        descr = load_descr(f"{dataset}.md", descr_module=descr_module)
+        print(descr)
+        print(json.dumps(descr, indent=2))
+        # Load data format
+        the_data_format = DATASETS[dataset]["format"]
+        format_ext = FORMATS[the_data_format]["ext"]
+        data_loader = FORMATS[the_data_format]["loader"]
+        # Infer dataset path
+        dataset_path = resources.files(data_module).joinpath(f"{dataset}.{format_ext}")
+
+        if not os.path.exists(dataset_path) or overwrite:
+            # Parse the url from the description file
+            url_content = descr[url_key]
+            if not is_url(url_content):
+                # Try to extract URL from markdown link format [text](url)
+                match = re.search(r"\[([^\]]+)\]\(([^)]+)\)", url_content)
+                url_path = match.group(2) if match else None
+            else:
+                url_path = url_content
+            assert is_url(
+                url_path
[…]
