### 2025-10-07 5e55a29
style: pre-commit fixes
