+++ py-phylo2vec/phylo2vec/datasets/descr/m501.md
@@ -1,8 +1,8 @@
 # M501 dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
-| -------------- | -------------- | --------------- | ---- |------------ | ---- |
-| Acanthocephalans | 29             | 2520         | RNA  | 18S rRNA sequences of Acanthocephala (thorny-headed worms) and Rotifera species. Originally retrieved from [treebase](https://www.treebase.org). | [FASTA link](https://bitbucket.org/XavMeyer/coevrj/raw/fa1355af4c28123bcf06fdb0bc43c90595beca45/data/adaptiveTreeProp/alignments/TreeBase/M2131.fasta) |
+| Taxonomic Rank   | Number of Taxa | Number of Bases | Type | Description                                                                                                                                      | Link                                                                                                                                                   |
+| ---------------- | -------------- | --------------- | ---- | ------------------------------------------------------------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| Acanthocephalans | 29             | 2520            | RNA  | 18S rRNA sequences of Acanthocephala (thorny-headed worms) and Rotifera species. Originally retrieved from [treebase](https://www.treebase.org). | [FASTA link](https://bitbucket.org/XavMeyer/coevrj/raw/fa1355af4c28123bcf06fdb0bc43c90595beca45/data/adaptiveTreeProp/alignments/TreeBase/M2131.fasta) |
 
 References:
 
