+++ py-phylo2vec/phylo2vec/datasets/descr/h3n2.md
@@ -1,8 +1,8 @@
 # Influenza A dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
-| -------------- | -------------- | --------------- | ---- |------------ | ---- |
-| Influenza A subtype H3N2 | 19   | 1407           | RNA  | Haemagglutinin nucleotide sequences isolated between 2011 and 2013. Retrieved from [treetime](https://github.com/neherlab/treetime_examples). | [FASTA Link](https://raw.githubusercontent.com/neherlab/treetime_examples/79eae7f8025a8ef3165c856c7359d92e738eb893/data/h3n2_na/h3n2_na_20.fasta) |
+| Taxonomic Rank           | Number of Taxa | Number of Bases | Type | Description                                                                                                                                   | Link                                                                                                                                              |
+| ------------------------ | -------------- | --------------- | ---- | --------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------- |
+| Influenza A subtype H3N2 | 19             | 1407            | RNA  | Haemagglutinin nucleotide sequences isolated between 2011 and 2013. Retrieved from [treetime](https://github.com/neherlab/treetime_examples). | [FASTA Link](https://raw.githubusercontent.com/neherlab/treetime_examples/79eae7f8025a8ef3165c856c7359d92e738eb893/data/h3n2_na/h3n2_na_20.fasta) |
 
 References:
 
