+++ py-phylo2vec/phylo2vec/datasets/descr/fluA.md
@@ -1,7 +1,7 @@
 # Influenza A dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
-| -------------- | -------------- | --------------- | ---- |------------ | ---- |
+| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description                                                                                                                                                              | Link                                                                                                                           |
+| -------------- | -------------- | --------------- | ---- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------ |
 | Influenza A    | 69             | 987             | RNA  | Haemagglutinin nucleotide sequences isolated between 1981 and 1998. Retrieved from [bv-brc](https://www.bv-brc.org) via [phylostan](https://github.com/4ment/phylostan). | [FASTA Link](https://raw.githubusercontent.com/4ment/phylostan/2ac68ac87cc6057efe556fa178fa391607ec5c3b/examples/fluA/fluA.fa) |
 
 References:
