+++ py-phylo2vec/phylo2vec/datasets/descr/yeast.md
@@ -1,8 +1,8 @@
 # Yeast dataset
 
-| Taxonomic Rank | Number of Taxa | Number of Bases | Type | Description | Link |
-| -------------- | -------------- | --------------- | ---- |------------ | ---- |
-| Fungi; mainly _Saccharomyces_ | 8              | 127026         | DNA  | Alignment of 106 orthologous genes of 8 different species of yeast. Retrieved from [phangorn](https://github.com/KlausVigo/phangorn/). | [RData link](https://github.com/KlausVigo/phangorn/raw/refs/heads/main/data/yeast.RData) |
+| Taxonomic Rank                | Number of Taxa | Number of Bases | Type | Description                                                                                                                            | Link                                                                                     |
+| ----------------------------- | -------------- | --------------- | ---- | -------------------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------- |
+| Fungi; mainly _Saccharomyces_ | 8              | 127026          | DNA  | Alignment of 106 orthologous genes of 8 different species of yeast. Retrieved from [phangorn](https://github.com/KlausVigo/phangorn/). | [RData link](https://github.com/KlausVigo/phangorn/raw/refs/heads/main/data/yeast.RData) |
 
 ## Download
 
