### 2025-10-08 98d8d5d
Doc data (#131)
* docs(data): add download links for each dataset + data section in readme

* data(py): add alignment loading for penn2023 datasets + tests + docs

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
