+++ py-phylo2vec/phylo2vec/datasets/__init__.py
@@ -1,5 +1,5 @@
 """Utilities to load/read/open datasets"""
 
-from ._base import load_descr, load_fasta, read_fasta
+from ._base import list_datasets, load_alignment, load_descr, load_fasta, read_fasta
 
-__all__ = ["load_descr", "load_fasta", "read_fasta"]
+__all__ = ["list_datasets", "load_alignment", "load_descr", "load_fasta", "read_fasta"]
