+++ py-phylo2vec/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "py-phylo2vec"
 # Python and binding version
-version = "1.3.1"
+version = "1.5.0"
 edition.workspace = true
 authors.workspace = true
 description.workspace = true
