### 2025-10-10 686cefc
test(rs): add more edge-case tests
