### 2025-10-10 00bfbf2
style: pre-commit fixes
