### 2025-10-12 8a0c94b
test(rs): add more edge-case tests (#133)
* test(rs): add more edge-case tests

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
