### 2025-10-12 052dc97
docs(paper): check + add missing ORCID
