### 2025-10-12 3787234
docs(paper): check + add missing ORCID (#134)
