### 2025-10-13 6bea3a5
docs: update citations + format middle names (#135)
