+++ joss/paper.md
@@ -16,22 +16,22 @@ authors:
     orcid: 0009-0008-1790-597X
     affiliation: 2
     equal-contrib: true
-  - name: Matthew J Penn
+  - name: Matthew J. Penn
     orcid: 0000-0001-8682-5393
     affiliation: 1
-  - name: Anthony MV Jakob
+  - name: Anthony M.V. Jakob
     orcid: 0000-0002-0996-1356
     affiliation: 4
-  - name: Frederik Mølkjær Andersen
+  - name: Frederik M. Andersen
     orcid: 0009-0004-4071-3707
     affiliation: 1
-  - name: Mark P Khurana
+  - name: Mark P. Khurana
     orcid: 0000-0002-1123-7674
     affiliation: 1
   - name: Landung Setiawan
     orcid: 0000-0002-1624-2667
     affiliation: 2
-  - name: David A Duchêne
+  - name: David A. Duchêne
     orcid: 0000-0002-5479-1974
     affiliation: 1
   - name: Samir Bhatt
