+++ CITATION.cff
@@ -7,25 +7,33 @@ authors:
   - family-names: "Nag"
     given-names: "Ayush"
     orcid: "https://orcid.org/0009-0008-1790-597X"
+  - family-names: "Penn"
+    given-names: "Matthew J"
+    orcid: "https://orcid.org/0000-0001-8682-5393"
+  - family-names: "Jakob"
+    given-names: "Anthony MV"
+    orcid: "https://orcid.org/0000-0002-0996-1356"
+  - family-names: "Andersen"
+    given-names: "Frederik M"
+    orcid: "https://orcid.org/0009-0004-4071-3707"
+  - family-names: "Khurana"
+    given-names: "Mark P"
+    orcid: "0000-0002-1123-7674"
   - family-names: "Setiawan"
     given-names: "Landung"
     orcid: "https://orcid.org/0000-0002-1624-2667"
-  - family-names: "Gordon"
-    given-names: "Madeline"
-    orcid: "https://orcid.org/0009-0003-6220-7218"
-title: "Phylo2Vec: a vector representation for binary trees"
+  - family-names: "Duchêne"
+    given-names: "David A"
+    orcid: "https://orcid.org/0000-0002-5479-1974"
+  - family-names: "Bhatt"
+    given-names: "Samir"
+    orcid: "https://orcid.org/0000-0002-0891-4611"
+title: "phylo2vec: a library for vector-based phylogenetic tree manipulation"
 url: "https://phylo2vec.readthedocs.io"
 repository-code: "https://github.com/sbhattlab/phylo2vec"
-# -------------------------------
-# Zenodo DOI will be added after the first release
-# doi: 10.xxxx/zenodo.xxxxxxx
-# -------------------------------
-# Note for `version` and `date-released` meta below
-# Pointing to a specific release/version will
-# require having to update this file with each release
-# version: x.x.x
-# date-released: xxxx-xx-xx
-# -------------------------------
+version: 1.5.0
+date-released: 2025-10-08
+doi: 10.5281/zenodo.17295653
 preferred-citation:
   type: article
   authors:
@@ -42,7 +50,9 @@ preferred-citation:
   - family-names: "Bhatt"
     given-names: "Samir"
   journal: "Systematic Biology"
-  doi: https://doi.org/10.1093/sysbio/syae030
+  doi: 10.1093/sysbio/syae030
   title: "Phylo2Vec: a vector representation for binary trees"
+  volume: 74
+  issue: 2
   month: 03
   year: 2025
