### 2025-10-14 b4af1fe
docs(paper): minor edits
