+++ joss/paper.bib
@@ -265,7 +265,6 @@
 @book{felsenstein2004,
   title={Inferring phylogenies},
   author={Felsenstein, Joseph},
-  volume={2},
   year={2004},
   publisher={Sinauer Associates},
   address={Sunderland, MA},
