### 2025-10-14 sbhattlab/phylo2vec#137 @Neclow
docs(paper): Editorial suggestions
https://github.com/openjournals/joss-reviews/issues/9040

* Ensure author names match ORCID names in draft + Zenodo
* Line 47: change "New features include" to something like "New features in latest release include"?
* Line 48: "leaf-level operations (pruning, placement, MRCA identification)" - Consider adding "and" before the last item in the list for proper serial comma usage
* Line 64: "Read The Docs" should be "Read the Docs"
