### 2025-10-14 0c78302
docs(paper): make author names match ORCID
