### 2025-10-14 81b81c1
docs(paper): update penn2024 citation
