### 2025-10-14 cea1f9d
docs(paper): update pdf
