+++ joss/paper.bib
@@ -2,8 +2,10 @@
   title={Phylo2Vec: a vector representation for binary trees},
   author={Penn, Matthew J and Scheidwasser, Neil and Khurana, Mark P and Duch{\^e}ne, David A and Donnelly, Christl A and Bhatt, Samir},
   journal={Syst. Biol.},
-  pages={syae030},
-  year={2024},
+  volume = {74},
+  number = {2},
+  pages = {250-266},
+  year = {2025},
   doi={10.1093/sysbio/syae030},
   publisher={Oxford University Press US}
 }
