### 2025-10-14 2ac6d4b
docs(paper): editorial suggestions (#138)
* docs(paper): minor edits

* docs(paper): make author names match ORCID

* docs(paper): update pdf

* docs(paper): update penn2024 citation

* docs(paper): update pdf
