+++ CITATION.cff
@@ -11,13 +11,13 @@ authors:
     given-names: "Matthew J"
     orcid: "https://orcid.org/0000-0001-8682-5393"
   - family-names: "Jakob"
-    given-names: "Anthony MV"
+    given-names: "Anthony"
     orcid: "https://orcid.org/0000-0002-0996-1356"
   - family-names: "Andersen"
-    given-names: "Frederik M"
+    given-names: "Frederik Mølkjær"
     orcid: "https://orcid.org/0009-0004-4071-3707"
   - family-names: "Khurana"
-    given-names: "Mark P"
+    given-names: "Mark Poulsen"
     orcid: "0000-0002-1123-7674"
   - family-names: "Setiawan"
     given-names: "Landung"
