### 2025-10-16 #9040 @abhishektiwari
@Neclow

Line 21: Spelling consistency for British English "utilize" should be "utilise"
Line 43: "addresses optimisations limitations" should be "addresses optimisation limitations" or "addresses limitations of optimisations"

**Reference formatting issues**: Can you please check each reference is matching original citation? A large number of them seems incorrectly formatted. Maybe this is your editor auto-correcting but they should match original paper title. If you click through DOI you can see differences.

Line 80: should be "Curious Parallels and Curious Connections—Phylogenetic Thinking in Biology and Historical Linguistic"
Line 87: should be "BEAGLE: An Application Programming Interface and High-Performance Computing Library for Statistical Phylogenetics "
Line 96: should be "AGC: compact representation of assembled genomes with fast queries and updates"
Line 99: should be "RecPhyloXML: a format for reconciled gene trees"
Line 103: should be "Inferring Phylogenies"
Line 115: should be "Rust-Bio: a fast and safe bioinformatics library"
Line 121: should be "Numba: a LLVM-based Python JIT compiler"
Line 134: should be "ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R"
Line 140: should be "Leaping through Tree Space: Continuous Phylogenetic Inference for Rooted and Unrooted Trees"
Line 145: should be "Chronumental: time tree estimation from very large phylogenies"
Line 151: should be "Ultrafast Sample placement on Existing tRees (UShER) enables real-time phylogenetics for the SARS-CoV-2 pandemic"
Line 55: should be "testthat: Get Started with Testing"
Line 157: should be "Molecular Evolution: A Statistical Approach"
