### 2025-10-16 d7c20ad
docs(paper): more editorial fixes
