### 2025-10-16 4e1ec05
docs(paper): more editorial fixes (#139)
