+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @article{penn2024,
-  title={Phylo2Vec: a vector representation for binary trees},
+  title={{Phylo2Vec: A Vector Representation for Binary Trees}},
   author={Penn, Matthew J and Scheidwasser, Neil and Khurana, Mark P and Duch{\^e}ne, David A and Donnelly, Christl A and Bhatt, Samir},
   journal={Syst. Biol.},
   volume = {74},
@@ -11,8 +11,8 @@
 }
 
 @article{paradis2019,
-    title = {ape 5.0: an environment for modern phylogenetics and
-      evolutionary analyses in {R}},
+    title = {{ape 5.0: an environment for modern phylogenetics and
+      evolutionary analyses in R}},
     author = {Emmanuel Paradis and Klaus Schliep},
     journal = {Bioinformatics},
     year = {2019},
@@ -34,7 +34,7 @@
 }
 
 @article{ayres2012,
-  title={BEAGLE: an application programming interface and high-performance computing library for statistical phylogenetics},
+  title={{BEAGLE: An Application Programming Interface and High-Performance Computing Library for Statistical Phylogenetics}},
   author={Ayres, Daniel L and Darling, Aaron and Zwickl, Derrick J and Beerli, Peter and Holder, Mark T and Lewis, Paul O and Huelsenbeck, John P and Ronquist, Fredrik and Swofford, David L and Cummings, Michael P and others},
   journal={Syst. Biol.},
   volume={61},
@@ -52,11 +52,11 @@
   volume={146},
   pages={263--266},
   year={1962},
-  note={In Russian. English translation in Soviet Mathematics Doklady, 3:1259--1263}
+  note={In Russian. English translation in Soviet Mathematics - Doklady, 3:1259--1263}
 }
 
 @article{penn2023,
-  title={Leaping through tree space: continuous phylogenetic inference for rooted and unrooted trees},
+  title={{Leaping through Tree Space: Continuous Phylogenetic Inference for Rooted and Unrooted Trees}},
   author={Penn, Matthew J and Scheidwasser, Neil and Penn, Joseph and Donnelly, Christl A and Duch{\^e}ne, David A and Bhatt, Samir},
   journal={Genome Biol. Evol.},
   volume={15},
@@ -69,7 +69,7 @@
 
 @article{duchemin2018,
     author = {Duchemin, Wandrille and Gence, Guillaume and Arigon Chifolleau, Anne-Muriel and Arvestad, Lars and Bansal, Mukul S and Berry, Vincent and Boussau, Bastien and Chevenet, Fran{\c{c}}ois and Comte, Nicolas and Dav{\'\i}n, Adri{\'a}n A and others},
-    title = {RecPhyloXML: a format for reconciled gene trees},
+    title = {{RecPhyloXML: a format for reconciled gene trees}},
     journal = {Bioinformatics},
     volume = {34},
     number = {21},
@@ -93,7 +93,7 @@
 }
 
 @article{koster2016,
-  title={Rust-Bio: a fast and safe bioinformatics library},
+  title={{Rust-Bio: a fast and safe bioinformatics library}},
   author={K{\"o}ster, Johannes},
   journal={Bioinformatics},
   volume={32},
@@ -105,7 +105,7 @@
 }
 
 @article{huerta2016,
-  title={ETE 3: reconstruction, analysis, and visualization of phylogenomic data},
+  title={{ETE 3: Reconstruction, Analysis, and Visualization of Phylogenomic Data}},
   author={Huerta-Cepas, Jaime and Serra, Fran{\c{c}}ois and Bork, Peer},
   journal={Mol. Biol. Evol.},
   volume={33},
[…]
