### 2025-10-16 2f35db1
docs(paper): add missing issues and page numbers
