+++ joss/paper.bib
@@ -17,6 +17,7 @@
     journal = {Bioinformatics},
     year = {2019},
     volume = {35},
+    number = {3},
     pages = {526-528},
     doi = {10.1093/bioinformatics/bty633},
 }
@@ -50,6 +51,7 @@
   author={Adelson-Velsky, Georgii, and Landis, Evgenii},
   journal={Proc. USSR Acad. Sci.},
   volume={146},
+  number={2},
   pages={263--266},
   year={1962},
   note={In Russian. English translation in Soviet Mathematics - Doklady, 3:1259--1263}
@@ -73,7 +75,7 @@
     journal = {Bioinformatics},
     volume = {34},
     number = {21},
-    pages = {3646-3652},
+    pages = {3646--3652},
     year = {2018},
     month = {05},
     issn = {1367-4803},
@@ -124,7 +126,7 @@
   volume = {9},
   number = {101},
   pages = {6943},
-  author = {Matthew Andres Moreno and Mark T. Holder and Jeet Sukumaran},
+  author = {Moreno, Matthew Andres and Holder, Mark T. and Sukumaran, Jeet},
   title = {{DendroPy 5: a mature Python library for phylogenetic computing}},
   journal = {J. Open Source Softw.}
 }
@@ -146,6 +148,7 @@
   author={Sanderson, Theo},
   journal={eLife},
   volume={11},
+  pages = {e82392},
   year={2022},
   doi={10.7554/eLife.82392},
   publisher={eLife Sciences Publications, Ltd}
