### 2025-10-16 a96e21e
docs(paper): fix some BE typos + oxford commas
