### 2025-10-16 b74dce8
docs(paper): fix author names in ola paper
