+++ joss/paper.bib
@@ -225,7 +225,7 @@
 
 @misc{richman2025,
   title={Vector encoding of phylogenetic trees by ordered leaf attachment},
-  author={Harry Richman and Cheng Zhang and Frederick A. Matsen IV},
+  author={Richman, Harry and Zhang, Cheng and Matsen IV, Frederick A},
   year={2025},
   journal={arXiv},
   eprint={2503.10169},
