### 2025-10-18 13312f9
chore: update min. rust version to 1.82 (#143)
