+++ pixi.toml
@@ -54,12 +54,12 @@ solve-group = "default"
 
 # ====== Default Dependencies ======
 [dependencies]
-rust = ">=1.81.0,<2"
+rust = ">=1.82.0,<2"
 openssl = ">=3.3.1,<3.4"
 pkg-config = ">=0.29.2,<0.30"
 compilers = ">=1.7.0,<1.8"
 clang = ">=16.0.6,<17"
-rust-src = ">=1.81.0,<2"
+rust-src = ">=1.82.0,<2"
 pre-commit = ">=4.0.1,<5"
 
 [target.linux-64.dependencies]
