### 2025-10-18 49c93da
chore(py): allow numpy<2.3
