+++ py-phylo2vec/pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
     "Operating System :: Microsoft :: Windows",
 ]
 dependencies = [
-  "numpy>=1.22,<2.1",
+  "numpy>=1.22,<2.3",
   "biopython>=1.84.0",
   "joblib>=1.2.0",
   "ete4==4.4.0",
