### 2025-10-18 ff343e1
feat(rs): use ryu float-to-string conversion
