### 2025-10-19 88eeb6d
chore(rs): update profiling and benchmarking scripts
