### 2025-10-19 8fe1440
chore(rs): pass dtype as arg1 in profiling
