### 2025-10-19 ee493e8
chore(rs): make profile comamnd default + update profiling docs
