+++ pixi.toml
@@ -29,7 +29,7 @@ start-release = "cargo run --release"
 open-doc = "cargo doc --open"
 benchmark = "cargo bench && open target/criterion/report/index.html"
 install-samply = "cargo install samply"
-profile = { cmd = "cargo build --profile profiling && samply record target/profiling/profile_main vector 100000", depends-on = ["install-samply"] }
+profile = { cmd = "cargo build --profile profiling && samply record target/profiling/profile_main", depends-on = ["install-samply"] }
 
 # ======= CI Environments ==========
 [environments]
