### 2025-10-19 b0d0162
style: pre-commit fixes
