### 2025-10-19 c96139e
chore(rs): update profiling and benchmarking scripts (#146)
* chore(rs): update profiling and benchmarking scripts

* chore(rs): pass dtype as arg1 in profiling

* chore(rs): make profile comamnd default + update profiling docs

* style: pre-commit fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
