### 2025-10-19 #9040 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
